### 2021-06-25 #3405 @whedon
[PRE REVIEW]: bfit: A Python Application For Beta-Detected NMR
**Submitting author:** @dfujim (<a href="http://orcid.org/0000-0003-2847-2053">Derek Fujimoto</a>)
**Repository:** <a href="https://github.com/dfujim/bfit" target ="_blank">https://github.com/dfujim/bfit</a>
**Version:** v4.9.1
