### 2021-06-25 #3405 @kyleniemeyer
In the meantime, it looks like your paper ran into some compilation errors. I think you might need to remove the comments from the file, I don't know if our parser can handle those.
