### 2021-07-15 #3405 @kthyng
Fair enough @pdebuyl. I think I saw another submission you'd be a good fit for in which case I'll invite you there. You could put up an [ooo](https://joss.readthedocs.io/en/latest/editing.html#out-of-office) for the time you'll be out if you'd like to, also.
