### 2021-07-16 #3405 @rkurchin
Willing to give it a go, but I don't really know anything about NMR or that particular characterization community, so I wouldn't be able to do a great job of suggesting reviewers if obvious candidates from the list can't do it. I'll leave it up to you, feel free to assign me if you think it's the best fit.
