### 2021-07-22 fb47c1b
Fixed issue with viewing 1n runs with no ppg param logged
