+++ bfit/__init__.py
@@ -12,7 +12,7 @@ from pkg_resources import parse_version
 from tkinter import messagebox
 
 __all__ = ['gui','fitting','backend','test']
-__version__ = '4.9.2'
+__version__ = '4.9.3'
 __author__ = 'Derek Fujimoto'
 logger_name = 'bfit'
 icon_path = os.path.join(os.path.dirname(__file__),'data','icon.gif')
