### 2021-07-23 907099a
Rates show as k/M/G instead of full number, added F/B
