### 2021-07-23 aa7cfa5
Fixed custom epics var issue
