### 2021-07-23 430ad62
Added 1c, 1d modes to raw scan calc list
