### 2021-07-23 a70e3ba
Fixed zerodiv error, area detection in fileview
