### 2021-07-23 93d8b27
More elegant fix for the area.upper issue
