### 2021-07-23 7e65376
removed 1n from fitting options
