### 2021-07-23 ade904b
Now accesses custom field of asym for drawing custom epics var
