### 2021-07-24 #3405 @danielskatz
👋 @lucydot - Would this be something you could edit as your first JOSS editing assignment?
