### 2021-07-24 #3405 @danielskatz
If so, you can use `@whedon assign me as editor` to assign yourself, or let us know if not
