### 2021-07-24 2645f59
Fixed area check in dataline
