### 2021-07-24 775ff4a
Fixed window_title set and get depreciation
