### 2021-07-24 eb2a1a8
turns off update autoscale only for 20 and 2h runs
