### 2021-07-25 038eed6
Reorded do_end_of_fit
