### 2021-07-25 4f54ccc
Fixed some tests, save shift_peak0
