### 2021-07-25 4f0b6ea
Fixed issue with not removing parameters from fitdata.fitpar on changing n comp
