### 2021-07-25 c62680a
Added legend to function placer
