### 2021-07-26 #3405 @lucydot
Hi @dfujim I'll be your editor for this submission 👋

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list ](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822)of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
