### 2021-08-02 #3405 @dfujim
@lucydot Standouts from the end of the list include villaa (Nuclear Physics, Particle Physics) and benjaminbolling (materials chemistry, physics).

Our experiment is somewhat unique in that we are investigating material properties with an NMR probe, but we do so with particle/nuclear physics methods. The technique is similar to muon spin rotation, but much less common. Apart from our lab, I think only [ISOLDE](https://isolde.cern/) is actively conducting implanted-ion beta-NMR experiments.

That being said, I don't think you actually need all too much topic-specific knowledge to review the software. Since our experiments cover a wide range of topics (superconductivity, soft matter, biological processes, battery materials, heterostructures, etc), the software is very general purpose to accommodate each experiment and is mostly concerned with basic data analysis.
