### 2021-08-04 2a61d94
Turned off text wrap in fileview
