+++ bfit/gui/tab_fileviewer.py
@@ -485,30 +485,6 @@ class fileviewer(object):
         except AttributeError:
             pass
         
-        try:
-            data_se["Doppler Tube CH0"] = "%.3f +/- %.3f V" % \
-                    (np.around(data.epics.dopplertube_ch0.mean, 3), 
-                     np.around(data.epics.dopplertube_ch0.std, 3))
-            key_order_se.append('Doppler Tube CH0')
-        except AttributeError:
-            pass
-        
-        try:
-            data_se["Doppler Tube CH1"] = "%.3f +/- %.3f V" % \
-                    (np.around(data.epics.dopplertube_ch1.mean, 3), 
-                     np.around(data.epics.dopplertube_ch1.std, 3))
-            key_order_se.append('Doppler Tube CH1')
-        except AttributeError:
-            pass
-        
-        try:
-            data_se["Doppler Tube CH2"] = "%.3f +/- %.3f V" % \
-                    (np.around(data.epics.dopplertube_ch2.mean, 3), 
-                     np.around(data.epics.dopplertube_ch2.std, 3))
-            key_order_se.append('Doppler Tube CH2')
-        except AttributeError:
-            pass
-        
         # get data: beam energy
         try: 
             init_bias = data.epics.target_bias.mean
@@ -539,6 +515,35 @@ class fileviewer(object):
         
         if key_order_se[-1] != '':
             key_order_se.append('')
+        
+        # doppler tubes
+        try:
+            data_se["Doppler Tube CH0"] = "%.3f +/- %.3f V" % \
+                    (np.around(data.epics.dopplertube_ch0.mean, 3), 
+                     np.around(data.epics.dopplertube_ch0.std, 3))
+            key_order_se.append('Doppler Tube CH0')
+        except AttributeError:
+            pass
+        
+        try:
+            data_se["Doppler Tube CH1"] = "%.3f +/- %.3f V" % \
+                    (np.around(data.epics.dopplertube_ch1.mean, 3), 
+                     np.around(data.epics.dopplertube_ch1.std, 3))
+            key_order_se.append('Doppler Tube CH1')
+        except AttributeError:
+            pass
+        
+        try:
+            data_se["Doppler Tube CH2"] = "%.3f +/- %.3f V" % \
+                    (np.around(data.epics.dopplertube_ch2.mean, 3), 
+                     np.around(data.epics.dopplertube_ch2.std, 3))
+            key_order_se.append('Doppler Tube CH2')
+        except AttributeError:
+            pass
+        
+        if key_order_se[-1] != '':
+            key_order_se.append('')
+        
         
         # laser stuff
         try: 
