### 2021-08-06 451d505
Fixed fwhm=0 issue in init par gui
