### 2021-08-06 129e436
Fixed export issue not using NBM when requested
