+++ bfit/gui/bfit.py
@@ -1426,7 +1426,7 @@ class bfit(object):
         
         # get asymmetry
         asym = data.asym(hist_select=self.hist_select, omit=omit, 
-                         rebin=rebin)
+                         rebin=rebin, nbm=self.use_nbm.get())
         
         # get new keys
         asym_out = {}
@@ -1454,6 +1454,7 @@ class bfit(object):
                     '# Rebin: %d' % rebin, 
                     '# Bin omission string: "%s"' % omit, 
                     '# Alternate histogram selection string: "%s"' % self.hist_select, 
+                    '# Data computed from NBM: %s' % self.use_nbm.get(),
                     '# ', 
                     '# Generated by bfit v%s on %s' % (__version__, datetime.datetime.now()), 
                     '#']
