### 2021-08-07 b4b5dde
Added neg height initial param
