### 2021-08-08 628f957
Fix for fitting no hist error
