### 2021-08-10 7792ce5
Added T1 to drawing options
