### 2021-08-10 df3895f
Added HH6 calculator, fixed calcs for other values
