+++ bfit/gui/calculator_nqr_B0_hh6.py
@@ -11,11 +11,11 @@ import bfit
 from bfit import logger_name
 
 # =========================================================================== #
-class calculator_nqr_B0(object):
+class calculator_nqr_B0_hh6(object):
     
     # ======================================================================= #
     def __init__(self, commandline=False):
-        """Draw window for Zaher's calculator"""
+        """Draw window for HH6 bfield calculator"""
         
         # get logger
         self.logger = logging.getLogger(logger_name)
@@ -25,7 +25,7 @@ class calculator_nqr_B0(object):
         if commandline: root = Tk()
         else:           root = Toplevel()
         
-        root.title("Zaher's Calculator")
+        root.title("HH6 Bfield Calculator")
         root.columnconfigure(0, weight=1)
         root.rowconfigure(0, weight=1)
 
@@ -50,7 +50,7 @@ class calculator_nqr_B0(object):
         
         # Entry and other objects
         title_line = ttk.Label(mainframe,   
-                text='BNQR Magnetic Static Field -- Current Converter\nFor ILE2A1:HH3 (the older one)', 
+                text='BNQR Magnetic Static Field -- Current Converter\nFor ILE2A1:HH6 (the newer one)', 
                 justify=CENTER)
         self.entry_field = Entry(mainframe, textvariable=self.field, width=10, 
                 justify=RIGHT)
@@ -60,19 +60,20 @@ class calculator_nqr_B0(object):
                 textvariable=self.current, width=10, justify=RIGHT)
         amperes = ttk.Label(mainframe, text='Amperes')
                 
-        link = ttk.Button(mainframe, 
-                          text='Calibration Data Here', 
-                          command=lambda : webbrowser.open(\
-                                  'http://bnmr.triumf.ca/?file=HH_Calibration'))
+        notes_line = ttk.Label(mainframe,   
+                            text='Inferred from runs 45019 and 45020 (2021) '+\
+                                 '\nwith NMR at 2.2 T\nA better calibration is '+\
+                                 'likely warrented',
+                            justify=CENTER)
         
         # Gridding
-        title_line.grid(        column=0, row=0, padx=5, pady=5, columnspan=5)
+        title_line.grid(        column=0, row=0, padx=5, pady=5, columnspan=6)
         self.entry_field.grid(  column=0, row=1, padx=5, pady=5)
         gauss.grid(             column=1, row=1, padx=5, pady=5)
         equals.grid(            column=2, row=1, padx=20, pady=5)
         self.entry_current.grid(column=3, row=1, padx=5, pady=5)
         amperes.grid(           column=4, row=1, padx=5, pady=5)
-        link.grid(              column=0, row=3, padx=5, pady=5, columnspan=5)
+        notes_line.grid(        column=0, row=3, padx=5, pady=5, columnspan=6)
         
         self.root = root
         
@@ -100,7 +101,7 @@ class calculator_nqr_B0(object):
                 field = float(self.field.get()) 
                 value = field2current(field)
                 self.current.set("%.4f" % np.around(value, 4))
-                self.logger.debug('Field of %g G converted to current of %g A (HH3)', 
+                self.logger.debug('Field of %g G converted to current of %g A (HH6)', 
                                  field, value)
             except ValueError:
                 self.current.set('')
@@ -111,12 +112,12 @@ class calculator_nqr_B0(object):
                 current = float(self.current.get()) 
                 value = current2field(current)
                 self.field.set("%.4f" % np.around(value, 4))
-                self.logger.debug('Current of %g A converted to field of %g G (HH3)', 
+                self.logger.debug('Current of %g A converted to field of %g G (HH6)', 
                                  current, value)
             except ValueError:
                 self.field.set('debug')
             
[…]
