### 2021-08-10 22da910
inspect: added NQR fields to epics window
