### 2021-08-10 4a66dcc
Added extra data point to HH6 calc
