### 2021-08-10 #3405 @jpata
> @jpata would you be able to review given your experience with data pipelines for large experimental collaborations?

@lucydot I can review this paper, I will look at it this week.
