### 2021-08-10 #3598 @whedon
[REVIEW]: bfit: A Python Application For Beta-Detected NMR
**Submitting author:** @dfujim (<a href="http://orcid.org/0000-0003-2847-2053">Derek Fujimoto</a>)
**Repository:** <a href="https://github.com/dfujim/bfit" target ="_blank">https://github.com/dfujim/bfit</a>
**Version:** 4.14.1

@nicksisco1932 @jpata, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

- [x] #3600

- [x] #3599
