### 2021-08-11 #3598 @lucydot
@jpata @nicksisco1932 a quick pointer on the JOSS review process. It is best to raise any issues during the review process in the software repository (https://github.com/dfujim/bfit) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below.
