### 2021-08-23 a4194fa
Moved draw command to fitdata obj
