+++ bfit/backend/fitdata.py
@@ -9,6 +9,8 @@ from bfit.gui.calculator_nqr_B0 import current2field
 from bfit.gui.calculator_nqr_B0_hh6 import current2field as current2field_hh6
 from bfit import logger_name
 
+from bfit.backend.raise_window import raise_window
+
 import numpy as np
 import pandas as pd
 
@@ -40,7 +42,7 @@ class fitdata(object):
                         Columns are fit_files.fitinputtab.collist
             id:         key for unique idenfication (str)    
             label:      label for drawing (StringVar)
-            mode:       run mode (str)
+            mode:       run mode (str, ex: 1f)
             omit:       omit bins, 1f only (StringVar)
             parnames:   parameter names in the order needed by the fit function
             rebin:      rebin factor (IntVar)
@@ -50,14 +52,14 @@ class fitdata(object):
     """
      
     # ======================================================================= #
-    def __init__(self, parentbfit, bd):
+    def __init__(self, bfit, bd):
         
         # get logger
         self.logger = logging.getLogger(logger_name)
         self.logger.debug('Initializing run %d (%d).', bd.run, bd.year)
         
-        # top level pointer
-        self.bfit = parentbfit
+        # top level pointers
+        self.bfit = bfit
         
         # bdata access
         self.bd = bd
@@ -68,6 +70,11 @@ class fitdata(object):
         self.label = StringVar()
         self.check_state = BooleanVar()
         
+        self.nbm = self.bfit.use_nbm
+        
+        self.rebin.set(1)
+        self.omit.set('')
+        self.label.set('')
         self.check_state.set(False)
         
         # key for IDing file 
@@ -111,6 +118,664 @@ class fitdata(object):
             messagebox.showerror(title=type(err).__name__, message=str(err))
             self.logger.exception(str(err))
             raise err from None
+
+    # ======================================================================= #
+    def draw(self, asym_type, figstyle='', **drawargs):
+        """
+            Draw the selected file
+            
+            bfit:       bfit object
+            asym_type:  input for asymmetry calculation
+            figstyle:   figure style. One of "data", "fit", or "param"
+            drawargs:   passed to errorbar
+        """
+        
+        self.logger.info('Drawing run %d (%d). mode: %s, rebin: %d, '+\
+                     'asym_type: %s, style: %s, %s', 
+                     self.run, 
+                     self.year, 
+                     self.mode, 
+                     self.rebin.get(), 
+                     asym_type, 
+                     self.bfit.draw_style.get(), 
+                     drawargs)
+        
+        # useful pointers
+        bfit = self.bfit
+        plt = self.bfit.plt
+        rebin = self.rebin.get()
+        omit = self.omit.get()
+        
+        # check omit values
+        if omit == self.bfit.fetch_files.bin_remove_starter_line:
+            omit = ''
+        
+        # convert asym type
+        asym_type = self.bfit.asym_dict.get(asym_type, asym_type)
+        
+        # default label value
+        if 'label' not in drawargs.keys():
+            label = str(self.run)
+        else:
+            label = drawargs.pop('label', None)
+            
+        # set drawing style arguments
+        for k in bfit.style:
+            if k not in drawargs.keys():
+                drawargs[k] = bfit.style[k]
+        
+        # get drawing style (ex: stack)
+        draw_style = bfit.draw_style.get()
+        
+        # make new window
+        if draw_style == 'new' or not plt.active[figstyle]:
+            plt.figure(figstyle)
+        elif draw_style == 'redraw':
+            plt.clf(figstyle)
+            
+        ax = plt.gca(figstyle)
+        
+        # set axis offset
+        try:
+            ax.get_xaxis().get_major_formatter().set_useOffset(False)
+        except AttributeError:
+            pass
+        
[…]
