+++ bfit/gui/bfit.py
@@ -47,7 +47,6 @@ from bfit.gui.popup_set_ppm_reference import popup_set_ppm_reference
 from bfit.gui.popup_ongoing_process import popup_ongoing_process
 from bfit.gui.popup_set_histograms import popup_set_histograms
 from bfit.backend.PltTracker import PltTracker
-from bfit.backend.raise_window import raise_window
 from bfit.backend.fitdata import fitdata
 import bfit.backend.colors as colors
 
@@ -734,659 +733,7 @@ class bfit(object):
     
     # ======================================================================= #
     def do_save(self): self.fit_files.save_state()
-    
-    # ======================================================================= #
-    def draw(self, data, asym_type, rebin=1, option='', figstyle='', **drawargs):
-        """
-            Draw the selected file
-            
-            data:       bdata object
-            asym_type:  input for asymmetry calculation
-            rebin:      rebining on asymmetry calculation
-            option:     options passed to asymmetry calculation
-            figstyle:   figure style. One of "data", "fit", or "param"
-            drawargs:   passed to errorbar
-        """
         
-        self.logger.info('Drawing run %d (%d). mode: %s, rebin: %d, '+\
-                     'option: %s, style: %s, %s', 
-                     data.run, 
-                     data.year, 
-                     asym_type, 
-                     rebin, 
-                     option, 
-                     self.draw_style.get(), 
-                     drawargs)
-        
-        # get draw setting 
-        draw_style = self.draw_style
-        
-        # default label value
-        if 'label' not in drawargs.keys():
-            label = str(data.run)
-        else:
-            label = drawargs.pop('label', None)
-            
-        # set drawing style arguments
-        for k in self.style:
-            if k not in drawargs.keys():
-                drawargs[k] = self.style[k]
-        
-        # get drawing style
-        style = self.draw_style.get()
-        
-        # make new window
-        if style == 'new' or not self.plt.active[figstyle]:
-            self.plt.figure(figstyle)
-        elif style == 'redraw':
-            self.plt.clf(figstyle)
-            
-        ax = self.plt.gca(figstyle)
-        
-        # set axis offset
-        try:
-            ax.get_xaxis().get_major_formatter().set_useOffset(False)
-        except AttributeError:
-            pass
-        
-        # get asymmetry: raw scans
-        if asym_type == 'r' and data.mode in ['1f', '1n', '1w', '1c', '1d', '1x']:
-            a = data.asym('raw', omit=option, hist_select=self.hist_select, 
-                          nbm=self.use_nbm.get())
-            x = np.arange(len(a.p[0]))
-            idx_p = a.p[0]!=0
-            idx_n = a.n[0]!=0
-            
-            xlabel = 'Bin'
-            self.plt.errorbar(figstyle, data.id, x[idx_p], a.p[0][idx_p], a.p[1][idx_p], 
-                    label=label+"($+$)", **drawargs)
-            
-            self.plt.errorbar(figstyle, data.id, x[idx_n], a.n[0][idx_n], a.n[1][idx_n], 
-                    label=label+"($-$)", unique=False, **drawargs)
-            
-        # do 2e mode
-        elif data.mode == '2e':
-            
-            # get asym
-            a = data.asym(hist_select=self.hist_select)
-        
-            # draw
-            if asym_type in ["raw_c", "raw_h", "raw_hs"]:
-                
-                # make 3D axes
-                if type(self.plt.gcf(figstyle)) == type(None):   
-                    self.plt.figure(figstyle)
-                ax = self.plt.gcf(figstyle).add_subplot(111, projection='3d', 
-                                  label=str(len(self.plt.gcf(figstyle).axes)))
-                
-                # get rid of bad draw options
-                try:                del drawargs['capsize']
-                except KeyError:    pass
[…]
