### 2021-08-23 95ec323
Fixed normalized draw
