+++ bfit/gui/bfit.py
@@ -156,7 +156,7 @@ class bfit(object):
                             "Raw Scans", 
                             "Shifted Split", 
                             "Shifted Combined", 
-                            "Normalized Combined", 
+                            "Combined Normalized", 
                             "Histograms", 
                             "Positive Helicity", 
                             "Negative Helicity", 
@@ -166,7 +166,7 @@ class bfit(object):
                             "Raw Scans", 
                             "Shifted Split", 
                             "Shifted Combined", 
-                            "Normalized Combined", 
+                            "Combined Normalized", 
                             "Histograms", 
                             "Positive Helicity", 
                             "Negative Helicity", 
@@ -205,7 +205,7 @@ class bfit(object):
                             "Raw Scans", 
                             "Shifted Split", 
                             "Shifted Combined", 
-                            "Normalized Combined", 
+                            "Combined Normalized", 
                             "Histograms", 
                             "Positive Helicity", 
                             "Negative Helicity", 
@@ -222,7 +222,7 @@ class bfit(object):
                             ], 
                       '2h':["Combined Helicity", 
                             "Split Helicity", 
-                            "Normalized Combined", 
+                            "Combined Normalized", 
                             "Positive Helicity", 
                             "Alpha Diffusion", 
                             "Alpha Diff Normalized", 
@@ -247,8 +247,7 @@ class bfit(object):
                  "Matched Helicity"         :'hm', 
                  "Shifted Split"            :'hs', 
                  "Shifted Combined"         :'cs', 
-                 "Normalized Combined"      :'cn1', # mode 1
-                 "Combined Normalized"      :'cn2', # mode 2
+                 "Combined Normalized"      :'cn',
                  "Matched Peak Finding"     :'hp', 
                  "Raw Scans"                :'r', 
                  "Histograms"               :'rhist', 
