### 2021-08-26 f29fdba
Moved get_values to fitdata
