+++ bfit/backend/fitdata.py
@@ -33,6 +33,8 @@ class fitdata(object):
             bias_std:   platform bias in kV (float)
             check_state:(BooleanVar)  
             chi:        chisquared from fit (float)
+            dataline:   pointer to dataline object in fetch_files tab
+            fitline:    pointer to fitline object in fit_files tab
             drawarg:    drawing arguments for errorbars (dict)
             field:      magnetic field in T (float)
             field_std:  magnetic field standard deviation in T (float)
@@ -60,6 +62,8 @@ class fitdata(object):
         
         # top level pointers
         self.bfit = bfit
+        self.dataline = None
+        self.fitline = None
         
         # bdata access
         self.bd = bd
@@ -848,6 +852,227 @@ class fitdata(object):
         
         return (T, dT)
         
+        # ======================================================================= #
+    def get_values(self, select):
+        """ Get plottable values"""
+
+        # fit parameters
+        parnames = self.parnames
+        
+        # popup for adding new parameters
+        try:
+            pop_addpar = self.bfit.fit_files.pop_addpar
+        except AttributeError:
+            pop_addpar = {}
+        
+        # helper function
+        def fetch(obj, attr):
+            """
+                Try fetching, if fails, return np.nan
+            """
+            try:
+                if len(attr) == 1:
+                    return getattr(obj, attr[0])
+                else:
+                    obj = getattr(obj, attr[0])
+                    return fetch(obj, attr[1:])
+
+            except (KeyError, AttributeError):
+                return np.nan
+
+        # Data file options
+        if select == 'Temperature (K)':
+            val = fetch(self, ['temperature','mean'])
+            err = fetch(self, ['temperature', 'std'])
+
+        elif select == 'B0 Field (T)':
+            val = fetch(self, ['field'])
+            err = fetch(self, ['field_std'])
+
+        elif select == 'RF Level DAC':
+            try:
+                val = fetch(self, ['rf_dac','mean'])
+                err = fetch(self, ['rf_dac','std'])
+            except AttributeError:
+                pass
+
+        elif select == 'Platform Bias (kV)':
+            try:
+                val = fetch(self, ['bias'])
+                err = fetch(self, ['bias_std'])
+            except AttributeError:
+                pass
+
+        elif select == 'Impl. Energy (keV)':
+            val = fetch(self, ['beam_kev'])
+            err = fetch(self, ['beam_kev_err'])
+
+        elif select == 'Run Duration (s)':
+            val = fetch(self, ['duration'])
+            err = np.nan
+
+        elif select == 'Run Number':
+            val = fetch(self, ['run'])
+            err = np.nan
+
+        elif select == 'Sample':
+            val = fetch(self, ['sample'])
+            err = np.nan
+
+        elif select == 'Start Time':
+            val = fetch(self, ['start_time'])
+            err = np.nan
+
+        elif select == 'Title':
+            val = fetch(self, ['title'])
+            err = np.nan
+
+        elif select == '1000/T (1/K)':
+            val = 1000/fetch(self, ['temperature', 'mean'])
+            err = 1000*fetch(self, ['temperature', 'std'])/\
+                        (fetch(self, ['temperature', 'mean'])**2)
+
+        elif select == 'Chi-Squared':
+            val = fetch(self, ['chi'])
+            err = np.nan
+
+        elif select == 'Year':
[…]
