+++ bfit/gui/tab_fit_files.py
@@ -1522,8 +1522,8 @@ class fit_files(object):
 
     # ======================================================================= #
     def get_values(self, select):
-        """ Get plottable values"""
-
+        """ Get plottable values from all runs"""
+        
         data = self.bfit.data
         dlines = self.bfit.fetch_files.data_lines
 
@@ -1532,237 +1532,17 @@ class fit_files(object):
         runs.sort()
 
         self.logger.debug('Fetching parameter %s', select)
-
-        # parameter names
-        parnames = self.fitter.gen_param_names(self.fit_function_title.get(),
-                                               self.n_component.get())
-
-        # helper functions
-        def fetch(obj, attr):
-            """
-                Try fetching, if fails, return np.nan
-            """
-            
-            # recursively find the value
-            try:
-                if len(attr) == 1:
-                    return getattr(obj, attr[0])
-                else:
-                    obj = getattr(obj, attr[0])
-                    return fetch(obj, attr[1:])
-
-            except (KeyError, AttributeError):
-                return np.nan
-
-        # Data file options
-        if select == 'Temperature (K)':
-            val = [fetch(data[r], ['temperature','mean']) for r in runs]
-            err = [fetch(data[r], ['temperature', 'std']) for r in runs]
-
-        elif select == 'B0 Field (T)':
-            val = [fetch(data[r], ['field']) for r in runs]
-            err = [fetch(data[r], ['field_std']) for r in runs]
-
-        elif select == 'RF Level DAC':
-            try:
-                val = [fetch(data[r], ['rf_dac','mean']) for r in runs]
-                err = [fetch(data[r], ['rf_dac','std']) for r in runs]
-            except AttributeError:
-                pass
-
-        elif select == 'Platform Bias (kV)':
-            try:
-                val = [fetch(data[r], ['bias']) for r in runs]
-                err = [fetch(data[r], ['bias_std']) for r in runs]
-            except AttributeError:
-                pass
-
-        elif select == 'Impl. Energy (keV)':
-            val =  [fetch(data[r], ['beam_kev']) for r in runs]
-            err =  [fetch(data[r], ['beam_kev_err']) for r in runs]
-
-        elif select == 'Run Duration (s)':
-            val = [fetch(data[r], ['duration']) for r in runs]
-            err = [np.nan for r in runs]
-
-        elif select == 'Run Number':
-            val = [fetch(data[r], ['run']) for r in runs]
-            err = [np.nan for r in runs]
-
-        elif select == 'Sample':
-            val = [fetch(data[r], ['sample']) for r in runs]
-            err = [np.nan for r in runs]
-
-        elif select == 'Start Time':
-            val = [fetch(data[r], ['start_time']) for r in runs]
-            err = [np.nan for r in runs]
-
-        elif select == 'Title':
-            val = [fetch(data[r], ['title']) for r in runs]
-            err = [np.nan for r in runs]
-
-        elif select == '1000/T (1/K)':
-            val = [1000/fetch(data[r], ['temperature', 'mean']) for r in runs]
-            err = [1000*fetch(data[r], ['temperature', 'std'])/\
-                        (fetch(data[r], ['temperature', 'mean'])**2) \
-                   for r in runs]
-
-        elif select == 'Chi-Squared':
-            val = [fetch(data[r], ['chi']) for r in runs]
-            err = [np.nan for r in runs]
-
-        elif select == 'Year':
-            val = [fetch(data[r], ['year']) for r in runs]
[…]
