### 2021-08-26 b869aa0
Moved draw_fit to fitdata
