+++ bfit/gui/tab_fit_files.py
@@ -95,7 +95,7 @@ class fit_files(object):
     mode = ""
     chi_threshold = 1.5 # threshold for red highlight on bad fits
     n_fitx_pts = 500    # number of points to draw in fitted curves
-
+    
     # ======================================================================= #
     def __init__(self, fit_data_tab, bfit):
 
@@ -1021,133 +1021,6 @@ class fit_files(object):
         self.plt.tight_layout(figstyle)
         self.plt.legend(figstyle)
 
-        raise_window()
-
-    # ======================================================================= #
-    def draw_fit(self, id, figstyle, unique=True, asym_mode=None, **drawargs):
-        """
-            Draw fit for a single run
-
-            id: id of run to draw fit of
-            figstyle: one of "data", "fit", or "param" to choose which figure
-                    to draw in
-            mode: from bfit.get_asym_mode, string of mode to draw
-         """
-
-        self.logger.info('Drawing fit for run %s. %s', id, drawargs)
-
-        # get data and fit results
-        data = self.bfit.data[id]
-        fit_par = data.fitpar.loc[data.parnames, 'res'].values
-
-        # get draw style
-        style = self.bfit.draw_style.get()
-
-        # label reset
-        if 'label' not in drawargs.keys():
-            drawargs['label'] = self.bfit.data[id].label.get()
-        drawargs['label'] += ' (fit)'
-        label = drawargs['label']
-
-        # set drawing style
-        draw_id = data.id
-
-        # make new window
-        if style == 'new' or not self.plt.active[figstyle]:
-            self.plt.figure(figstyle)
-        elif style == 'redraw':
-            self.plt.figure(figstyle)
-            self.plt.clf(figstyle)
-
-        # set drawing style arguments
-        for k in self.bfit.style:
-            if k not in drawargs.keys() \
-                    and 'marker' not in k \
-                    and k not in ['elinewidth', 'capsize']:
-                drawargs[k] = self.bfit.style[k]
-
-        # linestyle reset
-        if drawargs['linestyle'] == 'None':
-            drawargs['linestyle'] = '-'
-
-        # draw
-        t, a, da = data.asym('c')
-
-        fitx = np.linspace(min(t), max(t), self.n_fitx_pts)
-
-        if self.mode in self.bfit.units:
-            unit = self.bfit.units[self.mode]
-            xlabel = self.bfit.xlabel_dict[self.mode] % unit[1]
-        else:
-            xlabel = self.bfit.xlabel_dict[self.mode]
-
-        # get fity
-        fity = data.fitfn(fitx, *fit_par)
-
-        # draw relative to peak 0
-        if self.bfit.draw_rel_peak0.get():
-            
-            # get reference
-            par = data.fitpar
-            
-            if 'peak_0' in par.index:   index = 'peak_0'
-            elif 'mean_0' in par.index: index = 'mean_0'
-            elif 'peak' in par.index:   index = 'peak'
-            elif 'mean' in par.index:   index = 'mean'
-            else:
-                msg = "No 'peak' or 'mean' fit parameter found. Fit with" +\
-                     " an appropriate function."
-                self.logger.exception(msg)
-                messagebox.error(msg)
-                raise RuntimeError(msg)
-            
-            ref = par.loc[index, 'res']
-            
-            # do the shift
-            fitx -= ref                    
-            fitx *= unit[0]
-            xlabel = 'Frequency Shift (%s)' % unit[1]
-            self.logger.info('Drawing as freq shift from peak_0')
-        
-        # ppm shift
-        elif self.bfit.draw_ppm.get():
-            
-            # check div zero
-            try:
[…]
