+++ bfit/gui/tab_fetch_files.py
@@ -1115,13 +1115,13 @@ class dataline(object):
             if self.check_data.get():
                 mode = self.bfit.draw_style.get()
                 self.bfit.draw_style.set('stack')
-                self.bfit.fit_files.draw_fit(id=self.id, unique=False, 
-                                             asym_mode=self.bfit.get_asym_mode(self.bfit.fetch_files), 
-                                             figstyle=figstyle)
+                self.bdfit.draw_fit(unique=False, 
+                                    asym_mode=self.bfit.get_asym_mode(self.bfit.fetch_files), 
+                                    figstyle=figstyle)
                 self.bfit.draw_style.set(mode)
             else:
-                self.bfit.fit_files.draw_fit(id=self.id, figstyle=figstyle, 
-                                             asym_mode=self.bfit.get_asym_mode(self.bfit.fetch_files))
+                self.bdfit.draw_fit(figstyle=figstyle, 
+                                    asym_mode=self.bfit.get_asym_mode(self.bfit.fetch_files))
                 
         # draw residual
         if self.check_res.get():
