### 2021-08-26 e9ad98a
Moved draw_residuals to fitdata
