+++ bfit/gui/tab_fit_files.py
@@ -936,94 +936,6 @@ class fit_files(object):
             self.fit_lines[k].populate(force_modify=True)
 
     # ======================================================================= #
-    def draw_residual(self, id, figstyle, rebin=1, **drawargs):
-        """Draw fitting residuals for a single run"""
-
-        self.logger.info('Drawing residual for run %s, rebin %d, '+\
-                         'standardized: %s, %s', id, rebin,
-                         self.bfit.draw_standardized_res.get(), drawargs)
-
-        # get draw setting
-        figstyle = 'data'
-        draw_style = self.bfit.draw_style
-        plt.ion()
-
-        # get data and fit results
-        data = self.bfit.data[id]
-        fit_par = data.fitpar.loc[data.parnames, 'res'].values
-        data = data.bd
-
-        # default label value
-        if 'label' not in drawargs.keys():
-            label = str(data.run)
-        else:
-            label = drawargs.pop('label', None)
-
-        # set drawing style arguments
-        for k in self.bfit.style:
-            if k not in drawargs.keys():
-                drawargs[k] = self.bfit.style[k]
-
-        # make new window
-        style = self.bfit.draw_style.get()
-        if style == 'new' or not self.plt.active[figstyle]:
-            self.plt.figure(figstyle)
-        elif style == 'redraw':
-            self.plt.clf(figstyle)
-
-        ax = self.plt.gca(figstyle)
-        ax.get_xaxis().get_major_formatter().set_useOffset(False)
-
-        # get draw style
-        style = self.bfit.draw_style.get()
-
-        # get residuals
-        x, a, da = data.asym(self.bfit.get_asym_mode(self), rebin=rebin)
-        res = a - data.fitfn(x, *fit_par)
-
-        # set x axis
-        if data.mode in self.bfit.units:
-            unit = self.bfit.units[data.mode]
-            x *= unit[0]
-            xlabel = self.bfit.xlabel_dict[self.mode] % unit[1]
-        else:
-            xlabel = self.bfit.xlabel_dict[self.mode]
-
-        # draw
-        if self.bfit.draw_standardized_res.get():
-            self.plt.errorbar(figstyle, id, x, res/da, np.zeros(len(x)),
-                              label=label, **drawargs)
-
-            # draw fill
-            ax = self.plt.gca(figstyle)
-            lim = ax.get_xlim()
-            for i in range(1, 4):
-                ax.fill_between(lim, -1*i, i, color='k', alpha=0.1)
-            self.plt.xlim(figstyle, lim)
-            self.plt.ylabel(figstyle, r'Standardized Residual ($\sigma$)')
-        else:
-            self.plt.errorbar(figstyle, id, x, res, da, label=label, **drawargs)
-            self.plt.ylabel(figstyle, 'Residual')
-
-        # draw pulse marker
-        if '2' in data.mode:
-            self.plt.axvline(figstyle, 'line', data.pulse_s, ls='--', color='k')
-            unq = False
-        else:
-            unq = True
-
-        # plot elements
-        self.plt.xlabel(figstyle, xlabel)
-        self.plt.axhline(figstyle, 'line', 0, color='k', linestyle='-', zorder=20,
-                        unique=unq)
-
-        # show
-        self.plt.tight_layout(figstyle)
-        self.plt.legend(figstyle)
-
-        raise_window()
-
-    # ======================================================================= #
     def draw_param(self, *args):
         """Draw the fit parameters"""
         figstyle = 'param'
