### 2021-08-26 da1611d
Renamed show_fn_composition to draw_fn_composition
