### 2021-08-26 7f6dd84
Moved save/load state to bfit.py
