+++ bfit/gui/bfit.py
@@ -19,7 +19,7 @@ except ImportError as errmsg:
     print('No 3D axes drawing available')
     print(errmsg)
 
-import sys, os, datetime, textwrap
+import sys, os, datetime, textwrap, yaml
 import webbrowser, subprocess, importlib, logging, warnings
 import numpy as np
 import pandas as pd
@@ -456,7 +456,7 @@ class bfit(object):
         root.bind("<Button-4>", self.scroll_binder) 
         root.bind("<Button-5>", self.scroll_binder)
         
-        root.bind("<Control-Key-o>", self.do_load)
+        root.bind("<Control-Key-o>", self.load_state)
         
         # event bindings
         root.protocol("WM_DELETE_WINDOW", self.on_closing)
@@ -494,8 +494,8 @@ class bfit(object):
         menu_file.add_command(label='Run Commands', command=lambda:popup_terminal(wref.proxy(self)))
         menu_file.add_command(label='Export Data', command=self.do_export)
         menu_file.add_command(label='Export Fits', command=self.do_export_fit)
-        menu_file.add_command(label='Save State', command=self.do_save)
-        menu_file.add_command(label='Load State', command=self.do_load)
+        menu_file.add_command(label='Save State', command=self.save_state)
+        menu_file.add_command(label='Load State', command=self.load_state)
         menu_file.add_command(label='Close All Figures', command=self.do_close_all)
         menu_file.add_command(label='Exit', command=sys.exit)
         menubar.add_cascade(menu=menu_file, label='File')
@@ -726,12 +726,6 @@ class bfit(object):
     
     # ======================================================================= #
     def do_export_fit(self): self.fit_files.export_fit()
-        
-    # ======================================================================= #
-    def do_load(self, *args): self.fit_files.load_state()
-    
-    # ======================================================================= #
-    def do_save(self): self.fit_files.save_state()
         
     # ======================================================================= #
     def draw_binder(self, *args):
@@ -984,6 +978,192 @@ class bfit(object):
         webbrowser.open('https://github.com/dfujim/bfit/wiki')
     
     # ======================================================================= #
+    def load_state(self, filename=None):
+        """
+            Load the state of the gui
+        """
+
+        # get the filename
+        if filename is None:
+            filename = filedialog.askopenfilename(filetypes=[('yaml', '*.yaml'),
+                                                             ('allfiles', '*')])
+            if not filename:
+                return
+
+        self.logger.info('Loading program state from %s', filename)
+
+        # load the object with the data
+        with open(filename, 'r') as fid:
+            from_file = yaml.safe_load(fid)
+
+        # clear loaded runs
+        fetch_tab = self.fetch_files
+        fetch_tab.remove_all()
+
+        # bfit parameters
+        self.style = from_file['style']
+        self.hist_select = from_file['hist_select']
+        self.draw_style.set(from_file['draw_style'])
+        self.draw_fit.set(from_file['draw_fit'])
+        
+        self.probe_species.set(from_file['probe_species'])
+        self.set_probe_species()
+        
+        self.minimizer.set(from_file['minimizer'])
+        self.set_fit_routine()
+        
+        self.norm_with_param.set(from_file['norm_with_param'])
+        self.draw_standardized_res.set(from_file['draw_standardized_res'])
+        self.use_nbm.set(from_file['use_nbm'])
+        self.draw_ppm.set(from_file['draw_ppm'])
+        self.draw_rel_peak0.set(from_file['draw_rel_peak0'])
[…]
