+++ bfit/gui/tab_fit_files.py
@@ -329,8 +329,8 @@ class fit_files(object):
 
         # save/load state -----------------------
         state_frame = ttk.Labelframe(fit_data_tab, text='Program State', pad=5)
-        state_save_button = ttk.Button(state_frame, text='Save', command=self.save_state)
-        state_load_button = ttk.Button(state_frame, text='Load', command=self.load_state)
+        state_save_button = ttk.Button(state_frame, text='Save', command=self.bfit.save_state)
+        state_load_button = ttk.Button(state_frame, text='Load', command=self.bfit.load_state)
 
         state_save_button.grid(column=1, row=0, padx=5, pady=5)
         state_load_button.grid(column=2, row=0, padx=5, pady=5)
@@ -1371,191 +1371,6 @@ class fit_files(object):
             self.input_enable_disable(child, state=state, first=False)
 
     # ======================================================================= #
-    def load_state(self, filename=None):
-        """
-            Load the state of the gui
-        """
-
-        # get the filename
-        if filename is None:
-            filename = filedialog.askopenfilename(filetypes=[('yaml', '*.yaml'),
-                                                             ('allfiles', '*')])
-            if not filename:
-                return
-
-        self.logger.info('Loading program state from %s', filename)
-
-        # load the object with the data
-        with open(filename, 'r') as fid:
-            from_file = yaml.safe_load(fid)
-
-        # clear loaded runs
-        fetch_tab = self.bfit.fetch_files
-        fetch_tab.remove_all()
-
-        # bfit parameters
-        self.bfit.style = from_file['style']
-        self.bfit.hist_select = from_file['hist_select']
-        self.bfit.draw_style.set(from_file['draw_style'])
-        self.bfit.draw_fit.set(from_file['draw_fit'])
-        
-        self.bfit.probe_species.set(from_file['probe_species'])
-        self.bfit.set_probe_species()
-        
-        self.bfit.minimizer.set(from_file['minimizer'])
-        self.bfit.set_fit_routine()
-        
-        self.bfit.norm_with_param.set(from_file['norm_with_param'])
-        self.bfit.draw_standardized_res.set(from_file['draw_standardized_res'])
-        self.bfit.use_nbm.set(from_file['use_nbm'])
-        self.bfit.draw_ppm.set(from_file['draw_ppm'])
-        self.bfit.draw_rel_peak0.set(from_file['draw_rel_peak0'])
-        self.bfit.thermo_channel.set(from_file['thermo_channel'])
-        self.bfit.units = from_file['units']
-        self.bfit.label_default.set(from_file['label_default'])
-        self.bfit.ppm_reference = from_file['ppm_reference']
-        self.bfit.update_period = from_file['update_period']
-        self.bfit.bnmr_data_dir = from_file['bnmr_data_dir']
-        self.bfit.bnqr_data_dir = from_file['bnqr_data_dir']
-        
-        # set deadtime correction
-        self.bfit.deadtime = from_file['deadtime']
-        self.bfit.deadtime_switch.set(from_file['deadtime_switch'])
-        self.bfit.deadtime_global.set(from_file['deadtime_global'])
-        
-        # fileviewer
-        fileviewer_tab = self.bfit.fileviewer
-        fileviewer_tab.year.set(from_file['fileview_year'])
-        fileviewer_tab.runn.set(from_file['fileview_run'])
-        fileviewer_tab.get_data()
-        fileviewer_tab.asym_type.set(from_file['fileview_asym_type'])
-        fileviewer_tab.rebin.set(from_file['fileview_rebin'])
-        fileviewer_tab.is_updating.set(from_file['fileview_is_updating'])
-
-        # fetch files
[…]
