### 2021-08-26 712004c
Fixed load state error with no fetched run in inspect tab
