### 2021-08-26 d38c95d
Added show citation
