+++ bfit/global_variables.py
@@ -1,0 +1,10 @@
+from mudpy import global_variables
+
+# Define some global variables needed for setup
+# Derek Fujimoto
+# Aug 2021
+
+from os.path import join
+
+__src__ = join('bfit', 'fitting')
+__version__ = '4.12.4'
