### 2021-08-27 2fed089
Added try except to using TkAgg
