### 2021-08-27 b61edbb
Changes needed for windows/unix install
