### 2021-08-31 #3598 @lucydot
Hi @nicksisco1932 how is your review going? I can see that @dfujim has responded to some of the points you raised [here](https://github.com/dfujim/bfit/commit/c335e7d8938663fe12c61afaceaaca3e1d081e8a). If you have any other issues to raise as part of the review process, open them up on the [repository issues page](https://github.com/dfujim/bfit/issues). It'd be  great to hear a summary of your thoughts on the software so far.

@jpata have you had a chance to start the review yet? Let me know if there is anything I can clarify regarding the review process.
