### 2021-08-31 ce5f269
Fixed omit bins
