### 2021-09-04 dfujim/bfit#105 @jpata
JOSS review: installation requirements
In the context of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/3598), I wanted to ask to clarify the installation instructions in the README or alternatively the requirements in setup.py.

On a fresh Python 3.8.10 virtual environment under Ubuntu 20.04

```
python3 -m venv bfit_env
source bfit_env/bin/activate
python3 -m pip install bfit
```

did not work due to missing Cython:
```
     command: /home/joosep/bfit_env/bin/python3 -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-qm7m_65s/bfit/setup.py'"'"'; __file__='"'"'/tmp/pip-install-qm7m_65s/bfit/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /tmp/pip-install-qm7m_65s/bfit/pip-egg-info
         cwd: /tmp/pip-install-qm7m_65s/bfit/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-qm7m_65s/bfit/setup.py", line 3, in <module>
        from Cython.Build import cythonize
    ModuleNotFoundError: No module named 'Cython'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```

After doing `python3 -m pip install cython`, it complains about numpy:

```
     command: /home/joosep/bfit_env/bin/python3 -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-8ppx9ejb/bfit/setup.py'"'"'; __file__='"'"'/tmp/pip-install-8ppx9ejb/bfit/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /tmp/pip-install-8ppx9ejb/bfit/pip-egg-info
         cwd: /tmp/pip-install-8ppx9ejb/bfit/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-8ppx9ejb/bfit/setup.py", line 4, in <module>
        import numpy, os
    ModuleNotFoundError: No module named 'numpy'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```

After doing additionally `python3 -m pip install numpy`, it complains about

```
  Building wheel for mud-py (setup.py) ... error
   command: /home/joosep/bfit_env/bin/python3 -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-nxrrvkv_/mud-py/setup.py'"'"'; __file__='"'"'/tmp/pip-install-nxrrvkv_/mud-py/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-zwwoo8jf
       cwd: /tmp/pip-install-nxrrvkv_/mud-py/
  Complete output (6 lines):
  usage: setup.py [global_opts] cmd1 [cmd1_opts] [cmd2 [cmd2_opts] ...]
     or: setup.py --help [cmd1 cmd2 ...]
     or: setup.py --help-commands
     or: setup.py cmd --help

  error: invalid command 'bdist_wheel'
  ERROR: Failed building wheel for mud-py
```

but the package install and loads successfully.

Am I missing something, or are the requirements not fully picked up from [setup.py](https://github.com/dfujim/bfit/blob/master/setup.py#L60)? In any case, perhaps this is something to be clarified in [README](https://github.com/dfujim/bfit#installation-and-use) to reduce friction on installation.
