### 2021-09-04 dfujim/bfit#106 @jpata
JOSS review: citation for FastNumericalIntegration
Given that externally-written source code such as [FastNumericalIntegration](https://github.com/dfujim/bfit/tree/master/bfit/fitting/FastNumericalIntegration_src) is packaged together with bfit and seems to be central to the functionality, it may be appropriate to also cite this software in the JOSS bfit paper.
