### 2021-09-04 #3598 @jpata
I have reviewed the package as of [v4.12.0](https://github.com/dfujim/bfit/tree/4.12.0) and the [paper](https://github.com/openjournals/joss-reviews/issues/3598#issuecomment-896177757).

In general, it looks like an impressive piece of work and suitable for JOSS in my opinion.

I have left two clarifying issues on the [bfit](https://github.com/dfujim/bfit) repo, and I have a minor question:
- [ ] https://github.com/dfujim/bfit/issues/105
- [ ] https://github.com/dfujim/bfit/issues/106
- [ ] In terms of specifics of large collaborations, it may be useful to acknowledge the relevant collaboration (TRIUMF) and software/analysis groups within the collaboration in case this software has profited from discussions and feedback internal to the collaboration.

After the above points are clarified/solved, I will be happy to recommend it for publication.

(apologies for the delay on my side due to vacation)
