### 2021-09-04 112f23c
Added wheel to pre-dependencies install
