+++ paper/paper.bib
@@ -104,3 +104,11 @@ url = {https://linkinghub.elsevier.com/retrieve/pii/S0926204015000181},
 volume = {68-69},
 year = {2015}
 }
+
+@misc{Cook2014,
+author = {Cook, J. D.},
+url = {https://www.codeproject.com/Articles/31550/Fast-Numerical-Integration},
+howpublished = {\url{https://www.codeproject.com/Articles/31550/Fast-Numerical-Integration}},
+year = {2014},
+note = {Retrieved 2019/11/11}
+}
