### 2021-09-04 30adae4
Added citation for fast num integration
