+++ paper/paper.bib
@@ -108,7 +108,6 @@ year = {2015}
 @misc{Cook2014,
 author = {Cook, J. D.},
 url = {https://www.codeproject.com/Articles/31550/Fast-Numerical-Integration},
-howpublished = {\url{https://www.codeproject.com/Articles/31550/Fast-Numerical-Integration}},
 year = {2014},
 note = {Retrieved 2019/11/11}
 }
