### 2021-09-04 dfujim/bfit#106 @dfujim
Added citation in 30adae4cfc799d803e3f634dddf51970c770bbc9

Thanks for the suggestion!
