### 2021-09-04 93ea58a
Removed accidental import of mudpy.global_variables
