+++ setup.py
@@ -1,7 +1,7 @@
 
 # install numpy and cython dependencies needed for the rest of the setup script
 from setuptools import dist
-dist.Distribution().fetch_build_eggs(['cython>=0.28', 'numpy>=1.19', 'wheel>=0.34'])
+dist.Distribution().fetch_build_eggs(['cython>=0.28', 'numpy>=1.19'])
 
 import setuptools
 from distutils.core import Extension
@@ -60,7 +60,8 @@ setuptools.setup(
     install_requires = ['cython >= 0.28', 'numpy >= 1.19', 'tqdm >= 4.25.0',
                       'bdata >= 6.6.6', 'matplotlib >= 2.2.4', 'pandas >= 0.23.0',
                       'pyyaml >= 5.1', 'scipy >= 1.2.0', 'iminuit >= 2.6.1', 
-                      'requests >= 2.25.0', 'argparse >= 1.4.0', 'pytest >= 4.5.0'],
+                      'requests >= 2.25.0', 'argparse >= 1.4.0', 'pytest >= 4.5.0',
+                      'wheel>=0.34'],
     package_data = {'': ['data']},
     entry_points = {'console_scripts':['bfit = bfit:main']},
     include_package_data = True,
