### 2021-09-04 dfujim/bfit#105 @dfujim
So the first two install issues were anticipated. I only recently learned how to get around this problem (see changes to setup.py in 0a5d81a222ef4b764821ec57a50cb46ebf055aaa) but these have not been released to PyPI just yet. In the [README on PyPI](https://pypi.org/project/bfit/4.12.4/) you can see the old install instructions which include notes that you need to pre-install Cython and numpy.

The wheel install error is new to me though. I've also been testing with Python 3.8.10 on Ubuntu 20.04, but I don't get this error and I'm not sure why. I've added `wheel` as a dependency which hopefully gets around the issue (9b067daa21ced13aab07bb57d5d31954c9460d13).

I'll put these changes in v4.12.5 shortly
