+++ bfit/global_variables.py
@@ -5,4 +5,4 @@
 from os.path import join
 
 __src__ = join('bfit', 'fitting')
-__version__ = '4.12.4'
+__version__ = '4.12.5'
