### 2021-09-05 c661273
Added variables from dataline to fitdata
