+++ bfit/gui/popup_prepare_data.py
@@ -1,0 +1,375 @@
+# Drawing style window
+# Derek Fujimoto
+# July 2018
+
+from tkinter import *
+from tkinter import ttk
+import numpy as np
+from bfit import logger_name
+from bfit.backend.entry_color_set import on_focusout, on_entry_click
+import bfit.backend.colors as colors
+import matplotlib as mpl
+import webbrowser
+import logging
+from functools import partial
+
+# ========================================================================== #
+class popup_prepare_data(object):
+    """
+        Popup window for preparing data for fitting. 
+        
+        Data fields:
+        
+        bfit:               bfit object
+        checkbutton_state:  Checkbutton, linked to bdfit.check_state
+
+        data:               fitdata object
+        entry_label:        ttk.Entry, set run label
+        label_flatten_base: ttk.Label for flattening the baseline
+        omit_scan:          BooleanVar, if true, omit last scan from asym calc
+        flatten_base:       IntVar, if >0, flatten the baseline using this many bins
+        flatten_over:       BooleanVar, if true, apply baseline overcorrection in flatten
+    """
+
+    # ====================================================================== #
+    def __init__(self, bfit, data):
+        """
+            bfit: bfit object
+            data: fitdata object
+        """
+        
+        # set variables
+        self.bfit = bfit    # fit_files pointer
+        self.data = data
+        
+        # get logger
+        self.logger = logging.getLogger(logger_name)
+        self.logger.info('Initializing')
+        
+        # make a new window
+        self.win = Toplevel(bfit.mainframe)
+        self.win.title('Prepare %s for fitting' % data.get_values('Unique Id')[0])
+        topframe = ttk.Frame(self.win, relief='sunken', pad=5)
+
+        # icon
+        self.bfit.set_icon(self.win)
+
+        # show file details ---------------------------------------------------
+        frame_details = ttk.Labelframe(topframe, pad=5, text='File Details')
+        
+        r = 0
+        keys = (('Unique Id', '', ''), 
+                ('Title', '', ''), 
+                ('Temperature (K)',    'T    = ', ' K'), 
+                ('B0 Field (T)',       'B0   = ', ' T'), 
+                ('Platform Bias (kV)', 'Bias = ', ' kV'), 
+                ('Run Duration (s)', 'Duration: ', ''))
+        for key, label, unit in keys:
+            val, err = data.get_values(key)
+            
+            # prep output string
+            if key == 'Run Duration (s)':
+                
+                text = '%s%d:%d' % (label, int(val/60), val%60)
+                
+            else:
+                if type(val) is not str:
+                    val = '%.3f' % val
+                
+                text = '%s%s%s' % (label, val, unit)
+            
+            # set label in window
+            label = ttk.Label(frame_details, text=text, justify=LEFT)
+            label.grid(column=0, row=r, sticky='nw')
+            r += 1
+        
+        
+        # run label
+        frame_label = ttk.Frame(frame_details)
+        label_label = ttk.Label(frame_label, text='Label:', justify=LEFT)
+        self.entry_label = Entry(frame_label, textvariable=self.data.label, width=18)
+        self.entry_label.bind('<KeyRelease>', self.ungray_label)
+        button_label = ttk.Button(frame_label, text='Reset to default', 
+                                    command=self.reset_label, pad=1)
+        
+        # inspect button
+        button_inspect = ttk.Button(frame_details, text='Inspect', 
+                                    command=self.inspect, pad=1)
+        
+        # grid 
+        label_label.grid(column=0, row=0, sticky='w')
+        self.entry_label.grid(column=1, row=0, sticky='w')
[…]
