### 2021-09-08 2143abc
Finished initial version of popup_prepare data
