+++ bfit/gui/popup_prepare_data.py
@@ -12,6 +12,7 @@ import matplotlib as mpl
 import webbrowser
 import logging
 from functools import partial
+import matplotlib.pyplot as plt
 
 # ========================================================================== #
 class popup_prepare_data(object):
@@ -101,7 +102,8 @@ class popup_prepare_data(object):
         self.entry_label.grid(column=1, row=0, sticky='w')
         button_label.grid(column=2, row=0, sticky='w', padx=5)
         frame_label.grid(column=0, row=r, sticky='new'); r+=1
-        button_inspect.grid(column=0, row=r, sticky='ne', padx=5, pady=5); r+= 1
+        button_inspect.grid(column=0, row=r, sticky='se', padx=5, pady=5)
+        frame_details.grid_rowconfigure(r, weight=1)
         
         # mode ----------------------------------------------------------------
         mode_text = self.bfit.fetch_files.runmode_relabel[self.data.mode]
@@ -118,6 +120,16 @@ class popup_prepare_data(object):
         self.config_checkbutton_state()
         self.checkbutton_state.grid(column=0, row=0)
         
+        # rebin ---------------------------------------------------------------
+        frame_rebin = ttk.Labelframe(topframe, text='Rebin', pad=5)
+        label_rebin = ttk.Label(frame_rebin, text='Rebin:')
+        spin_rebin = Spinbox(frame_rebin, from_=0, to=100,
+                             width=3,
+                             textvariable=self.data.rebin)
+        
+        label_rebin.grid(column=0, row=0, padx=5, pady=5)
+        spin_rebin.grid(column=1, row=0, padx=5, pady=5)
+        
         # scan repair ---------------------------------------------------------
         frame_scan = ttk.Labelframe(topframe, text='Scan Repair', pad=5)
         
@@ -129,15 +141,12 @@ class popup_prepare_data(object):
         
         # omit last scan
         self.omit_scan = BooleanVar()
-        self.omit_scan.set(False)
         checkbutton_omit_scan = ttk.Checkbutton(frame_scan, text='Omit final scan if incomplete', 
                 variable=self.omit_scan, onvalue=True, offvalue=False, pad=5,
                 command=self.set_bin_repair)
         
         # flatten baseline
         self.flatten_base = IntVar()
-        self.flatten_base.set(0)
-        
         self.label_flatten_base = ttk.Label(frame_scan, 
                                       text='Flatten baseline using N end bins. N = ')
         spin_flatten_base = Spinbox(frame_scan, from_=0, to=100000,
@@ -146,12 +155,6 @@ class popup_prepare_data(object):
         spin_flatten_base.bind("<KeyRelease>", self.set_bin_repair)
         spin_flatten_base.bind("<Leave>", self.set_bin_repair)
  
-        self.flatten_over = BooleanVar()
-        self.flatten_over.set(False)        
-        checkbutton_flatten_over = ttk.Checkbutton(frame_scan, text='Ensure combined scans are flattened', 
-                variable=self.flatten_over, onvalue=True, offvalue=False, pad=5,
-                command=self.set_bin_repair)
-                
         frame_draw_corrected = ttk.Labelframe(frame_scan, text='Draw Correction', pad=5)
         button_draw_raw_flat = ttk.Button(frame_draw_corrected, text='Raw', 
                                     command=lambda: self.draw_compare(draw_style='r'))
@@ -179,7 +182,7 @@ class popup_prepare_data(object):
         r = 0
         label_omit.grid(column=0, row=r, sticky='w')
         entry_omit.grid(column=1, row=r, sticky='w')
-        button_omit.grid(column=3, row=r, sticky='e'); r+=1
+        button_omit.grid(column=3, row=r, sticky='e', padx=10); r+=1
[…]
