+++ bfit/gui/tab_fetch_files.py
@@ -27,6 +27,7 @@ class fetch_files(object):
             
             asym_type: StringVar, drawing style
             bfit: pointer to parent class
+            base_bins = IntVar, number of bins to use as baseline on scan ends
             canvas_frame_id: id number of frame in canvas
             check_rebin: IntVar for handling rebin aspect of checkall
             check_bin_remove: StringVar for handing omission of 1F data
@@ -44,6 +45,7 @@ class fetch_files(object):
             fet_entry_frame: frame of fetch tab
             listbox_history: listbox for run input history
             max_number_fetched: max number of files you can fetch
+            omit_state: BooleanVar, if true set omit all final incomplete scans
             run: StringVar input to fetch runs.
             runmode_label: display run mode
             runmode: display run mode list of strings
@@ -153,31 +155,62 @@ class fetch_files(object):
         
         # Frame for group set options ----------------------------------------
         right_frame = ttk.Labelframe(bigright_frame, \
-                text='Operations on Checked Items', pad=30)
+                text='Operations on Active Items', pad=30)
         
         check_remove = ttk.Button(right_frame, text='Remove', \
                 command=self.remove_all, pad=5)
         check_draw = ttk.Button(right_frame, text='Draw', \
                 command=lambda:self.draw_all('data'), pad=5)
         
-        check_rebin_label = ttk.Label(right_frame, text="Rebin:", pad=5)
-        check_rebin_box = Spinbox(right_frame, from_=1, to=100, width=3, \
+        
+        # Scan repair options ------------------------------------------------
+        frame_scan = ttk.Labelframe(right_frame, \
+                text='Scan Repair Options', pad=5)
+        
+        check_rebin_label = ttk.Label(frame_scan, text="Rebin:", pad=5)
+        check_rebin_box = Spinbox(frame_scan, from_=1, to=100, width=3, \
                 textvariable=self.check_rebin, 
-                command=self.set_all)
-        check_bin_remove_entry = Entry(right_frame, \
+                command=lambda: self.set_all('rebin'))
+        
+        check_bin_remove_label = ttk.Label(frame_scan, text="Remove:", pad=5)
+        check_bin_remove_entry = Entry(frame_scan, \
                 textvariable=self.check_bin_remove, width=20)
         
+        self.omit_state = BooleanVar()
+        omit_checkbox = ttk.Checkbutton(frame_scan, 
+                text='Omit incomplete final scan', variable=self.omit_state, 
+                onvalue=True, offvalue=False, pad=5, 
+                command=lambda: self.set_all('omit_scan'))
+        self.omit_state.set(False)
+        
+        self.base_bins = IntVar()
+        base_label = ttk.Label(frame_scan, text="baseline bins", pad=5)
+        base_spinbox = Spinbox(frame_scan, from_=0, to=1000, width=4, \
+                               textvariable=self.base_bins, 
+                               command=lambda: self.set_all('base_bins'))
+        self.base_bins.set(0)
+        
+        # grid
+        check_rebin_label.grid(     column=0, row=0, sticky='e')
+        check_rebin_box.grid(       column=1, row=0, sticky='w')
+        check_bin_remove_label.grid(column=0, row=1, sticky='e')
+        check_bin_remove_entry.grid(column=1, row=1, sticky='w', columnspan=2)
+        omit_checkbox.grid(column=0, row=2, sticky='w', columnspan=2)
+        base_spinbox.grid(column=0, row=3, sticky='e')
+        base_label.grid(column=1, row=3, sticky='w')
+        
         # key bindings
-        check_bin_remove_entry.bind('<KeyRelease>', self.set_all)
-        check_rebin_box.bind('<KeyRelease>', self.set_all)
[…]
