+++ bfit/backend/fitdata.py
@@ -27,6 +27,7 @@ class fitdata(object):
         
         Data Fields:
             
+            base_bins:  n bins to use in baseline flatten (IntVar)
             bd:         bdata object for data and asymmetry (bdata)
             bfit:       pointer to top level parent object (bfit)
             bias:       platform bias in kV (float)
@@ -35,7 +36,7 @@ class fitdata(object):
             check_draw_data: BooleanVar, draw data?
             check_draw_fit: BooleanVar, draw fit?
             check_draw_res: BooleanVar, draw residuals?
-            check_state: BooleanVar
+            check_state: BooleanVar, include in fit?
             
             chi:        chisquared from fit (float)
             dataline:   pointer to dataline object in fetch_files tab
@@ -51,10 +52,10 @@ class fitdata(object):
             label:      label for drawing (StringVar)
             mode:       run mode (str, ex: 1f)
             omit:       omit bins, 1f only (StringVar)
+            omit_scan:  if true omit incomplete scan (BoolVar)
             parnames:   parameter names in the order needed by the fit function
             rebin:      rebin factor (IntVar)
             run:        run number (int)
-            scan_repair_options: string
             year:       run year (int)
               
     """
@@ -80,21 +81,26 @@ class fitdata(object):
         # fetch files variables
         self.check_state = BooleanVar()
         self.rebin = IntVar()
+        self.base_bins = IntVar()
         self.omit = StringVar()
         self.label = StringVar()
         self.check_draw_data = BooleanVar()
         self.check_draw_fit = BooleanVar()
         self.check_draw_res = BooleanVar()
+        self.omit_scan = BooleanVar()
 
         
         # fetch files defaults
         self.check_state.set(True)
         self.rebin.set(1)
+        self.base_bins.set(0)
         self.omit.set('')
         self.label.set('')
         self.check_draw_data.set(True)
         self.check_draw_fit.set(False)
         self.check_draw_res.set(False)
+        self.omit_scan.set(False)
+        
         self.scan_repair_options = ''
         self.parnames = []
         
@@ -136,8 +142,30 @@ class fitdata(object):
         
         # set repair options 
         if 'scan_repair_options' not in kwargs.keys():
-            kwargs['scan_repair_options'] = self.scan_repair_options
+            s1 = 'omit' if self.omit_scan.get() else ''
+            s2 = '%d' % self.base_bins.get()
+            kwargs['scan_repair_options'] = '%s:%s' % (s1, s2)
         
+        # rebin
+        if 'rebin' not in kwargs.keys():
+            kwargs['rebin'] = self.rebin.get()
+        
+        # omit
+        if 'omit' not in kwargs.keys():
+            omit = self.omit.get()
+            if omit == self.bfit.fetch_files.bin_remove_starter_line:
+                omit = ''
+        
+            kwargs['omit'] = omit
+            
+        # nbm
+        if 'nbm' not in kwargs.keys():
+            kwargs['nbm'] = self.nbm.get()
+                          
+        # hist select
+        if 'hist_select' not in kwargs.keys():
+            kwargs['hist_select'] = self.bfit.hist_select
+                          
         # check for errors
         try:
             return self.bd.asym(*args, deadtime=deadtime, 
@@ -148,7 +176,7 @@ class fitdata(object):
             raise err from None
 
     # ======================================================================= #
-    def draw(self, asym_type, figstyle='', **drawargs):
+    def draw(self, asym_type, figstyle='', asym_args=None, **drawargs):
         """
             Draw the selected file
             
@@ -168,15 +196,13 @@ class fitdata(object):
[…]

[…]
