### 2021-09-08 f3df173
Fixed error in draw_fit
