### 2021-09-10 b0e49f1
Moved keyvars to global_variables
