### 2021-09-10 8334645
Added test for filter
