### 2021-09-10 17d8553
Added error popup for filter
