### 2021-09-17 fc4c7cb
Filter instructions no longer blocks main window loop
