### 2021-09-18 284db60
Simplified fit lines. Not fully working
