+++ bfit/gui/tab_fit_files.py
@@ -22,6 +22,7 @@ from bfit.fitting.decay_31mg import fa_31Mg
 from bfit.fitting.functions import decay_corrected_fn
 from bfit.backend.entry_color_set import on_focusout, on_entry_click
 from bfit.backend.raise_window import raise_window
+from bfit.gui.InputLine import InputLine
 
 import numpy as np
 import pandas as pd
@@ -605,7 +606,7 @@ class fit_files(object):
 
         # reset modify all value
         self.set_as_group.set(modify_all_value)
-
+        
     # ======================================================================= #
     def do_add_param(self, *args):
         """Launch popup for adding user-defined parameters to draw"""
@@ -1370,21 +1371,24 @@ class fit_files(object):
             self.input_enable_disable(child, state=state, first=False)
 
     # ======================================================================= #
-    def modify_all(self, *args, source=None, par='', column=''):
+    def set_lines(self, pname, col, value):
         """
-            Modify all input fields of each line to match the altered one,
+            Modify all input fields of each line to match the altered one
             conditional on self.set_as_group
 
-            source_line: the fitline to copy
-            parameter:   name of the parameter to copy
-            column:      name of the column to copy
+            pname: string, parameter being changed
+            col:   str, column being changed
+            value: new value to assign
         """
-
-        setall = self.set_as_group.get()
-        self.logger.info('Set modify all as %s', setall)
-
-        for k in self.fit_lines.keys():
-            self.fit_lines[k].set_input(source, par, column, setall)
+    
+        for fitline in self.fit_lines.values():
+            
+            # get line id
+            id = [line.pname for line in fitline.lines].index(pname)
+            line = fitline.lines[id]
+            
+            # set 
+            line.set(**{col:value})        
 
     # ======================================================================= #
     def return_binder(self):
@@ -1518,10 +1522,8 @@ class fitline(object):
             dataline        pointer to dataline object in fetch_files_tab
             disable_entry_callback  disables copy of entry strings to
                                     dataline.bdfit parameter values
+            lines           list of InputLine objects
             parent          pointer to parent object (frame)
-            parlabels       label objects, saved for later destruction
-            parentry        [parname][colname] of Entry objects saved for
-                            retrieval and destruction
             run_label       label for showing which run is selected
             run_label_title label for showing which run is selected
             fitframe        mainframe for this tab.
@@ -1552,9 +1554,8 @@ class fitline(object):
         self.parent = parent
         self.dataline = dataline
         self.row = row
-        self.parlabels = []
-        self.parentry = {}
         self.disable_entry_callback = False
+        self.lines = []
 
         # get parent frame
         fitframe = ttk.Frame(self.parent, pad=(5, 0))
@@ -1588,17 +1589,17 @@ class fitline(object):
                         command=self.draw_fn_composition, pad=0)
 
         c = 0
-        ttk.Label(fitframe, text='Parameter').grid(    column=c, row=1, padx=5); c+=1
-        gui_param_button.grid(column=c, row=1, padx=5, pady=2); c+=1
-        ttk.Label(fitframe, text='Low Bound').grid(    column=c, row=1, padx=5); c+=1
-        ttk.Label(fitframe, text='High Bound').grid(   column=c, row=1, padx=5); c+=1
-        result_comp_button.grid(column=c, row=1, padx=5, pady=2, sticky=(E, W)); c+=1
-        ttk.Label(fitframe, text='Error (-)').grid(        column=c, row=1, padx=5); c+=1
[…]
