### 2021-09-18 5e2d373
Fitting works, modify all works, typing in  new values does not
