+++ bfit/gui/tab_fit_files.py
@@ -428,33 +428,6 @@ class fit_files(object):
                             )
 
     # ======================================================================= #
-    def _make_shared_var_dict(self):
-        """Make the dictionary to make sure all shared checkboxes are synched"""
-
-        # get parameter list
-        try:
-            parlst = [p for p in self.fitter.gen_param_names(
-                                                self.fit_function_title.get(),
-                                                self.n_component.get())]
-
-        # no paramteters: empty out the variable list
-        except KeyError:
-            share_var = {}
-
-        # make new shared list
-        else:
-            # re-initialize
-            share_var = {p:BooleanVar() for p in parlst}
-
-            # set to old values if they exist
-            for p in parlst:
-                if p in self.share_var.keys():
-                    share_var[p].set(self.share_var[p].get())
-
-        # save to object
-        self.share_var = share_var
-
-    # ======================================================================= #
     def canvas_scroll(self, event):
         """Scroll canvas with files selected."""
         if event.num == 4:
@@ -514,8 +487,6 @@ class fit_files(object):
             self.fit_runmode_label['text'] = ""
             self.mode = ""
 
-        # make shared_var dictionary
-        self._make_shared_var_dict()
 
         # delete unused fitline objects
         for k in list(self.fit_lines.keys()):       # iterate fit list
@@ -593,8 +564,6 @@ class fit_files(object):
         self.xaxis_combobox['values'] = [''] + parlst + lst
         self.yaxis_combobox['values'] = [''] + parlst + lst
         self.annotation_combobox['values'] = [''] + parlst + lst
-
-        self._make_shared_var_dict()
 
         # turn off modify all so we don't cause an infinite loop
         modify_all_value = self.set_as_group.get()
@@ -711,34 +680,11 @@ class fit_files(object):
             # bdata object
             bdfit = fitline.dataline.bdfit
 
-            # pdict
+            # get entry values
             pdict = {}
-            for parname in fitline.parentry.keys():
-
-                # get entry values
-                pline = fitline.parentry[parname]
-                line = []
-                for col in fitline.collist:
-
-                    # get number entries
-                    if col in ('p0', 'blo', 'bhi'):
-                        try:
-                            line.append(float(pline[col][0].get()))
-                        except ValueError as errmsg:
-                            self.logger.exception("Bad input.")
-                            messagebox.showerror("Error", str(errmsg))
-                            raise errmsg
-
-                    # get "Fixed" entry
-                    elif col in ['fixed']:
-                        line.append(pline[col][0].get())
-
-                    # get "Shared" entry
-                    elif col in ['shared']:
-                        line.append(pline[col][0].get())
-
-                # make dict
-                pdict[parname] = line
+            for line in fitline.lines:
+                inpt = line.get('*')
+                pdict[line.pname] = [inpt[k] for k in ('p0', 'blo', 'bhi', 'fixed', 'shared')]
 
             # doptions
             doptions = {}
@@ -746,12 +692,12 @@ class fit_files(object):
             if self.use_rebin.get():
                 doptions['rebin'] = bdfit.rebin.get()
 
-            if self.mode in ('1f', '1w', '1x'):
+            if '1' in self.mode:
                 dline = self.bfit.fetch_files.data_lines[key]
                 doptions['omit'] = dline.bin_remove.get()
                 if doptions['omit'] == dline.bin_remove_starter_line:
                     doptions['omit'] = ''
-            elif self.mode in ('20', '2h', '2e'):
+            elif '2' in self.mode:
                 pass
             else:
[…]
