+++ bfit/gui/tab_fit_files.py
@@ -502,7 +502,7 @@ class fit_files(object):
                 if k in self.fit_lines_old.keys():
                     self.fit_lines[k] = self.fit_lines_old[k]
                 else:
-                    self.fit_lines[k] = fitline(self.bfit, self.runframe, dl[k], n)
+                    self.fit_lines[k] = fitline(self.bfit, self.runframe, dl[k].bdfit, n)
             self.fit_lines[k].grid(n)
             n+=1
 
@@ -678,7 +678,7 @@ class fit_files(object):
             fitline = self.fit_lines[key]
 
             # bdata object
-            bdfit = fitline.dataline.bdfit
+            bdfit = fitline.data
 
             # get entry values
             pdict = {}
@@ -1119,7 +1119,7 @@ class fit_files(object):
         keylist = []
         for k, line in self.fit_lines.items():
             keylist.append(k)
-            data = line.dataline.bdfit
+            data = line.data
             
             for kk in data.fitpar.index:
                 
@@ -1316,7 +1316,7 @@ class fit_files(object):
             self.input_enable_disable(child, state=state, first=False)
 
     # ======================================================================= #
-    def set_lines(self, pname, col, value):
+    def set_lines(self, pname, col, value, skipline=None):
         """
             Modify all input fields of each line to match the altered one
             conditional on self.set_as_group
@@ -1324,6 +1324,7 @@ class fit_files(object):
             pname: string, parameter being changed
             col:   str, column being changed
             value: new value to assign
+            skipline: if this line, don't modify
         """
     
         for fitline in self.fit_lines.values():
@@ -1332,14 +1333,12 @@ class fit_files(object):
             id = [line.pname for line in fitline.lines].index(pname)
             line = fitline.lines[id]
             
+            if line == skipline:
+                continue
+            
             # set 
             line.set(**{col:value})     
-            
-            # change inital value
-            value = float(value)
-            fitpar = self.bfit.data[line.pname].fitpar
-            fitpar.loc[line.pname, col] = value
-
+                        
     # ======================================================================= #
     def return_binder(self):
         """
@@ -1469,7 +1468,7 @@ class fitline(object):
         Instance variables
 
             bfit            pointer to top class
-            dataline        pointer to dataline object in fetch_files_tab
+            data            fitdata object in bfit.data dictionary
             disable_entry_callback  disables copy of entry strings to
                                     dataline.bdfit parameter values
             lines           list of InputLine objects
@@ -1482,7 +1481,7 @@ class fitline(object):
     collist = ['p0', 'blo', 'bhi', 'res', 'dres-', 'dres+', 'chi', 'fixed', 'shared']
 
     # ======================================================================= #
-    def __init__(self, bfit, parent, dataline, row):
+    def __init__(self, bfit, parent, data, row):
         """
             Inputs:
                 bfit:       top level pointer
@@ -1495,12 +1494,12 @@ class fitline(object):
         # get logger
         self.logger = logging.getLogger(logger_name)
         self.logger.debug('Initializing fit line for run %d in row %d',
-                          dataline.run, row)
+                          data.run, row)
 
         # initialize
         self.bfit = bfit
         self.parent = parent
-        self.dataline = dataline
+        self.data = data
         self.row = row
         self.disable_entry_callback = False
         self.lines = []
@@ -1511,14 +1510,14 @@ class fitline(object):
         frame_title = ttk.Frame(fitframe)
 
         # label for displyaing run number
-        if type(self.dataline.bdfit.bd) is bdata:
+        if type(self.data.bd) is bdata:
             self.run_label = Label(frame_title,
-                            text='[ %d - %d ]' % (self.dataline.run,
[…]
