### 2021-09-20 be87f9e
Fixed linking to stringvar
