### 2021-09-20 d487765
Fixed setting new fns on tab change
