### 2021-09-20 4c60b57
Fixed rounding on set param from gui
