+++ bfit/gui/popup_param.py
@@ -26,6 +26,7 @@ class popup_param(object):
         fitter:         fit_tab.fitter obje (defined in default_routines.py)
         first:          if True, first time through fixing parameters
         fname:          name of the function 
+        lines:          dict of InputLine objects to link parameters
         logger:         logging variable
         mode:           1 or 2 to switch between run modes
         n_components:   number of components in the fit function
@@ -150,11 +151,12 @@ class popup_param(object):
                                           ncomp=self.n_components)
         
         lines = fit_tab.fit_lines[run_id].lines
+        self.lines = {line.pname:line  for line in lines}
         
         # make initial parameter list
-        self.p0 = {line.pname:line.variable['p0'] for line in lines}
-        self.blo = {line.pname:line.variable['blo'] for line in lines}
-        self.bhi = {line.pname:line.variable['bhi'] for line in lines}
+        self.p0 =  {k:v.variable['p0']  for k, v in self.lines.items()}
+        self.blo = {k:v.variable['blo'] for k, v in self.lines.items()}
+        self.bhi = {k:v.variable['bhi'] for k, v in self.lines.items()}
         
         # check for zero width
         for k in self.p0.keys():
@@ -287,7 +289,7 @@ class popup_param(object):
                         
             # set p0     
             val = p0[i][self.parmap[key]]
-            self.p0[k].set(val)
+            self.lines[k].set(p0=val)
                     
             # check bounds are ok
             blo = float(self.blo[k].get())
