### 2021-09-20 351b51c
Fixed assigning shared var
