### 2021-09-20 c8d4bbf
Fixed testing tab3_fit_files
