+++ bfit/test/test_tab3_fit_files.py
@@ -94,9 +94,9 @@ def separate_curve_fit(entry, run, **kwargs):
     # fit function
     pexp = pulsed_exp(lifetime = bd.life.Li8, pulse_len = dat.pulse_s)
 
-    p0 = [float(entry[k]['p0'][1].get()) for k in entry.keys()]
-    blo = [float(entry[k]['blo'][1].get()) for k in entry.keys()]
-    bhi = [float(entry[k]['bhi'][1].get()) for k in entry.keys()]
+    p0 = [float(entry[k]['p0'].get()) for k in entry.keys()]
+    blo = [float(entry[k]['blo'].get()) for k in entry.keys()]
+    bhi = [float(entry[k]['bhi'].get()) for k in entry.keys()]
     
     # fit
     t,a,da = dat.asym('c')
@@ -119,9 +119,9 @@ def separate_migrad(entry, run, do_minos=False, **kwargs):
     pexp = pulsed_exp(lifetime = bd.life.Li8, pulse_len = dat.pulse_s)
 
     # get p0, bounds
-    p0 = {k.replace('1_T1','lambda_s'):float(entry[k]['p0'][1].get()) for k in entry.keys()}
-    bounds = [( float(entry[k]['blo'][1].get()), 
-                float(entry[k]['bhi'][1].get())) for k in entry.keys()]
+    p0 = {k.replace('1_T1','lambda_s'):float(entry[k]['p0'].get()) for k in entry.keys()}
+    bounds = [( float(entry[k]['blo'].get()), 
+                float(entry[k]['bhi'].get())) for k in entry.keys()]
     
     # fit
     t,a,da = dat.asym('c')
@@ -164,15 +164,15 @@ def fit(separate_fit, minimizer, b=None, tab=None, tab2=None):
     
     for k, line in tab.fit_lines.items():
     
-        entry = line.parentry
+        entry = {l.pname: l.variable for l in line.lines}
         
         # get results separately
         par, lower, upper, chi2 = separate_fit(entry, k)    
         
         # get results displayed
-        res = [float(entry[k]['res'][1].get()) for k in entry.keys()]
-        low = [float(entry[k]['dres-'][1].get()) for k in entry.keys()]
-        upp = [float(entry[k]['dres+'][1].get()) for k in entry.keys()]
+        res = [float(entry[k]['res'].get()) for k in entry.keys()]
+        low = [float(entry[k]['dres-'].get()) for k in entry.keys()]
+        upp = [float(entry[k]['dres+'].get()) for k in entry.keys()]
         
         # check
         assert_array_almost_equal(par, res, err_msg = 'Copying %s minimizer results for %s' % (minimizer, k), decimal=3)
@@ -180,7 +180,7 @@ def fit(separate_fit, minimizer, b=None, tab=None, tab2=None):
         assert_array_almost_equal(upper, upp, err_msg = 'Copying %s minimizer upper errors for %s' % (minimizer, k), decimal=3)
         
         # check chi2
-        chi = float(entry['1_T1']['chi'][1].get())
+        chi = float(entry['1_T1']['chi'].get())
         assert_almost_equal(chi, chi2, 
                         err_msg = 'Copying chi2 for %s minimizer for %s' % (minimizer, k), 
                         decimal=2)
@@ -202,7 +202,7 @@ def fit_single(separate_fit, minimizer, b=None, tab=None, tab2=None):
 
     tab.populate()
     line = tab.fit_lines['2020.40123']
-    entry = line.parentry
+    entry = {l.pname: l.variable for l in line.lines}
     
     # set minimizer
     b.minimizer.set('bfit.fitting.fitter_%s' % minimizer)
[…]
