### 2021-09-20 eb66d74
Redefined template popup
