+++ bfit/gui/popup_fit_results.py
@@ -7,6 +7,7 @@ from tkinter import ttk, messagebox
 from functools import partial
 import logging
 import numpy as np
+import pandas as pd
 
 from bfit import logger_name
 from bfit.gui.template_fit_popup import template_fit_popup
@@ -20,6 +21,11 @@ class popup_fit_results(template_fit_popup):
         
         chi_label:      Label, chisquared output
         fittab:         notebook tab
+        new_par:        dataframe, index: parnames, columns: p0, blo, bhi, res, err
+        
+        output_par_text_val string, contents of output_par_text
+        output_text_val     dict of strings, contents of output_text
+       
         par:            list, fit results
         reserved_pars:  dict, keys: x, y vals: strings of parameter names
         
@@ -40,7 +46,67 @@ class popup_fit_results(template_fit_popup):
     def __init__(self, bfit, input_fn_text='', output_par_text='', output_text='', 
                  chi=np.nan, x='', y=''):
         
-        super().__init__(bfit, input_fn_text, output_par_text, output_text)
+        super().__init__(bfit, input_fn_text)
+        
+        self.output_par_text_val = output_par_text
+        
+        if not output_text:
+            self.output_text_val = {}
+        else:
+            self.output_text_val = output_text
+        
+        # text for output
+        output_frame = ttk.Frame(self.right_frame, relief='sunken', pad=5)
+        output_head1_label = ttk.Label(output_frame, text='Par Name')
+        output_head2_label = ttk.Label(output_frame, text='p0')
+        output_head3_label = ttk.Label(output_frame, text='Bounds')
+        output_head4_label = ttk.Label(output_frame, text='Result')
+        output_head5_label = ttk.Label(output_frame, text='Error (-)')
+        output_head6_label = ttk.Label(output_frame, text='Error (+)')
+        self.output_par_text = Text(output_frame, width=8, height=8)
+        self.output_text = {k:Text(output_frame, width=8, height=8, wrap='none')\
+                            for k in ('p0', 'blo', 'bhi', 'res', 'err-', 'err+')}
+        
+        # default starter strings
+        if self.output_par_text_val: 
+            self.output_par_text.insert('1.0', self.output_par_text_val)
+        self.output_par_text.config(state='disabled')
+    
+        if self.output_text_val: 
+            for k in self.output_text_val:
+                self.output_text[k].insert('1.0', self.output_text_val[k])
+
+        # disable results
+        for k in ('res', 'err-', 'err+'):
+            self.output_text[k].config(state='disabled', width=12)
+
+        # key bindings and scrollbar
+        scrollb_out = Scrollbar(output_frame, command=self.yview)
+        self.output_par_text['yscrollcommand'] = scrollb_out.set
+        for k in self.output_text:
+            self.output_text[k].bind('<KeyRelease>', self.get_result_input)
+            self.output_text[k]['yscrollcommand'] = scrollb_out.set
+                
+        c = 0; r = 0;
+        output_head1_label.grid(column=c, row=r);        c+=1;
+        output_head2_label.grid(column=c, row=r);        c+=1;
+        output_head3_label.grid(column=c, row=r, 
+                                columnspan=2);          c+=2;
+        output_head4_label.grid(column=c, row=r);        c+=1;
+        output_head5_label.grid(column=c, row=r);        c+=1;
+        output_head6_label.grid(column=c, row=r);        c+=1;
+        
+        c = 0; r += 1;
+        self.output_par_text.grid(column=c, row=r, sticky=N); c+=1;
+        for k in ('p0', 'blo', 'bhi', 'res', 'err-', 'err+'):
+            self.output_text[k].grid(column=c, row=r, sticky=N); c+=1;
+        
+        # fitting button 
[…]
