### 2021-09-20 3a8a199
Set up new GUI for constrains popup
