+++ bfit/gui/popup_fit_constraints.py
@@ -4,7 +4,7 @@
 
 
 from tkinter import *
-from tkinter import ttk
+from tkinter import ttk, messagebox
 from functools import partial
 
 import logging, re, os, warnings
@@ -24,11 +24,16 @@ import queue
 # ========================================================================== #
 class popup_fit_constraints(template_fit_popup):
     """
-        Popup window for modelling the fit results with a function
+        Popup window for settings fit parameters according to a function
         
         bfit
         fittab
         logger
+        
+        
+        label_new_var       Label, show new variables
+        label_defined       Label, show which variables will be redefined
+        
         
         output_par_text     text, detected parameter names
         output_text         dict, keys: p0, blo, bhi, res, err, value: tkk.Text objects
@@ -43,13 +48,12 @@ class popup_fit_constraints(template_fit_popup):
 
     # names of modules the constraints have access to
     modules = {'np':'numpy'}
-    
-    window_title = 'Fit data with contrained parameters'
+    window_title = 'Constrain parameters'
     
     # ====================================================================== #
-    def __init__(self, bfit, constr_text='', output_par_text='', output_text=''):
+    def __init__(self, bfit, input_fn_text=''):
         
-        super().__init__(bfit, constr_text, output_par_text, output_text)
+        super().__init__(bfit, input_fn_text)
         
         # Keyword parameters
         key_param_frame = ttk.Frame(self.left_frame, relief='sunken', pad=5)
@@ -90,18 +94,63 @@ class popup_fit_constraints(template_fit_popup):
         # Text entry
         self.entry_label['text'] = 'Enter one constraint equation per line.'+\
                                  '\nNon-reserved words are shared variables.'+\
-                                 '\nEx: "1_T1 = a*np.exp(b*BIAS**0.5)+c"'+\
-                                 '\nNote: Shared and fixed flags from main window ignored.'
+                                 '\nEx: "1_T1 = a*np.exp(b*BIAS**0.5)+c"'                                 
+                
+        # detected new parameters
+        frame_detected = ttk.Frame(self.right_frame)
+        label_defined_title = ttk.Label(frame_detected, text='Constrained parameters')
+        label_new_var_title = ttk.Label(frame_detected, text='New variables')
+        
+        self.label_new_var = ttk.Label(frame_detected, text='')
+        self.label_defined = ttk.Label(frame_detected, text='')
+                
+        # add constrain button
+        button_constrain = ttk.Button(self.right_frame, text='Constrain', 
+                                      command=self.set_constraints)
                 
         # gridding
         key_param_label.grid(column=0, row=0)
         fit_param_label.grid(column=0, row=0)
         modules_label.grid(column=0, row=0)
         
-        key_param_frame.grid(column=0, row=0, rowspan=1, sticky=(E, W), padx=1, pady=1)
-        module_frame.grid(column=0, row=1, sticky=(E, W), padx=1, pady=1, rowspan=2)
-        fit_param_frame.grid(column=0, row=3, sticky=(E, W, N, S), padx=1, pady=1)
+        key_param_frame.grid(column=0, row=0, rowspan=1, sticky='ew', padx=1, pady=1)
+        module_frame.grid(column=0, row=1, sticky='ew', padx=1, pady=1, rowspan=2)
+        fit_param_frame.grid(column=0, row=3, sticky='ewns', padx=1, pady=1)
         
+        frame_detected.grid(column=0, row=4, sticky='ewn', padx=1, pady=1)
+        frame_detected.grid_columnconfigure(0, weight=1)
+        frame_detected.grid_columnconfigure(1, weight=1)
+        label_defined_title.grid(column=0, row=0, sticky='n', padx=1, pady=1)
[…]
