### 2021-09-21 #3598 @lucydot
Hello @dfujim  - congratulations, we are very close to accepting `bfit` for publication 🎈

Please could you i) issue a new tagged release of the software (if changes have been made); ii) archive it (on, for example, Zenodo); iii) report the version number and archive DOI here on the thread.
