### 2021-09-21 #3598 @lucydot
@jpata @nicksisco1932 thank-you for reviewing for JOSS ✨
