### 2021-09-21 263ef06
Added input check, line disable/enable
