+++ bfit/gui/popup_fit_constraints.py
@@ -137,7 +137,12 @@ class popup_fit_constraints(template_fit_popup):
     def do_after_parse(self, defined=None, eqn=None, new_par=None):
         
         # show inputs as readback
-        s = '\n'.join(sorted(defined))
+        ncomp = self.bfit.fit_files.n_component.get()
+        fn_name = self.bfit.fit_files.fit_function_title.get()
+        par_names = self.bfit.fit_files.fitter.gen_param_names(fn_name, ncomp)
+        s = [d if d else '<blank>' for d in sorted(defined)]
+        s = [d if d in par_names else '%s [ERROR: Bad parameter]' % d for d in s]
+        s = '\n'.join(s)
         self.label_defined.config(text=s)
         
         s = sorted(np.unique(np.concatenate(new_par)))
@@ -262,114 +267,113 @@ class popup_fit_constraints(template_fit_popup):
         blo = np.array(blo).T
         bhi = np.array(bhi).T
         
-        # ~ # set up fitter inputs
-        # ~ npar = len(sharelist)
-        # ~ bounds = [[l, h] for l, h in zip(blo, bhi)]
-        # ~ data = [self.bfit.data[k] for k in keylist]
-        # ~ kwargs = {'p0':p0, 'bounds':bounds}
+        # set up fitter inputs
+        npar = len(sharelist)
+        bounds = [[l, h] for l, h in zip(blo, bhi)]
+        data = [self.bfit.data[k] for k in keylist]
+        kwargs = {'p0':p0, 'bounds':bounds}
         
-        # ~ # get minimizer
-        # ~ if 'trf'   in fit_files.fitter.__name__:  minimizer = 'trf'
-        # ~ if 'minos' in fit_files.fitter.__name__:  minimizer = 'minos'
-        # ~ if 'hesse' in fit_files.fitter.__name__:  minimizer = 'migrad'
+        # get minimizer
+        if 'trf'   in fit_files.fitter.__name__:  minimizer = 'trf'
+        if 'minos' in fit_files.fitter.__name__:  minimizer = 'minos'
+        if 'hesse' in fit_files.fitter.__name__:  minimizer = 'migrad'
         
-        # ~ # set up queue for results
-        # ~ que = Queue()
+        # set up queue for results
+        que = Queue()
         
-        # ~ # do fit
-        # ~ def run_fit():
-            # ~ try:
-                # ~ out = fit_bdata(data=data, 
-                                # ~ fn=fitfns, 
-                                # ~ shared=sharelist, 
-                                # ~ asym_mode='c', 
-                                # ~ rebin=rebin, 
-                                # ~ omit=omit, 
-                                # ~ xlims=None, 
-                                # ~ hist_select=self.bfit.hist_select, 
-                                # ~ minimizer=minimizer, 
-                                # ~ **kwargs)
-            # ~ except Exception as err:
-                # ~ que.put(str(err))
-                # ~ raise err from None
+        # do fit
+        def run_fit():
+            try:
+                out = fit_bdata(data=data, 
+                                fn=fitfns, 
+                                shared=sharelist, 
+                                asym_mode='c', 
+                                rebin=rebin, 
+                                omit=omit, 
+                                xlims=None, 
+                                hist_select=self.bfit.hist_select, 
+                                minimizer=minimizer, 
+                                **kwargs)
+            except Exception as err:
+                que.put(str(err))
+                raise err from None
                 
-            # ~ # par, std_l, std_u, cov, chi, gchi
-            # ~ que.put(out)
+            # par, std_l, std_u, cov, chi, gchi
+            que.put(out)
             
-        # ~ # start the fit
-        # ~ def do_enable():
-            # ~ fit_files.input_enable_disable(self.win, state='normal', first=False)
[…]
