### 2021-09-22 a8ff427
Fixed sharing constr
