### 2021-09-22 0b29317
Inputs calculated based on constrained eqns
