### 2021-09-22 e62cc22
Don't show template popup on init
