### 2021-09-22 0bcca8a
Removed uneeded checks for constraint popup
