### 2021-09-22 64a3f22
Fixed missing eqn check
