### 2021-09-27 #3598 @lucydot
Hi @dfujim apologies for the late reply. We have a preference for the Zenodo archive to be named the same as the JOSS paper (e.g., bfit: A Python Application For Beta-Detected NMR). Is it possible to rename accordingly? I'll then recommend acceptance to the editor-in-chiefs.
