### 2021-09-27 e018b50
Constrained fitting apparently working
