+++ bfit/fitting/fitter.py
@@ -246,8 +246,10 @@ class fitter(object):
             chis = [chis]
             
         output = {}
-        for i,d in enumerate(bdata_list):
+        for i, d in enumerate(bdata_list):
             key = self.keyfn(d)
+            # ~ print(keylist, pars, stds_l, stds_h, chis)
+
             output[key] = pd.DataFrame({'parname':keylist,
                                         'res': pars[i],
                                         'dres+': stds_h[i],
