### 2021-09-28 f6ef63f
Removed uneeded fitline in InputLine init
