### 2021-09-28 8eed46d
Fixed issue with constr param showing up after ncomp change
