### 2021-09-28 1e1851c
Removed do_fit from constr popup
