+++ bfit/gui/popup_fit_constraints.py
@@ -216,225 +216,6 @@ class popup_fit_constraints(template_fit_popup):
             self.new_par_unique = []
     
     # ====================================================================== #
-    def _do_fit(self, text):
-        """
-            Set up the fit functions and do the fit. Then map the outputs to the
-            proper displays. 
-        """
-        
-        self.logger.info('Starting fit')
-        
-        # get equations and defined variables
-        defined = [t.split('=')[0].strip() for t in text]
-        eqn = [t.split('=')[1].strip() for t in text]
-        
-        # check that the defined variables all match function inputs
-        for d in defined: 
-            if d not in self.parnames:
-                errmsg = 'Definition for "%s" invalid. ' % d+\
-                         'Must only define function inputs. '
-                messagebox.showerror("Error", errmsg)
-                raise RuntimeError(errmsg)
-        
-        # make shared parameters for the rest of the parameters
-        allpar = self.new_par['name'].tolist()
-        alldef = defined[:]     # all parameter names in order
-        sharelist = [True]*len(allpar)
-        
-        for n in sorted(self.parnames):
-            if n not in defined:
-                eqn.append(n)
-                alldef.append(n)
-                allpar.append(n)
-                sharelist.append(False)
-                        
-        # replace 1_T1 with lambda1
-        for i, _ in enumerate(allpar):
-            if '1_T1' in allpar[i]:
-                allpar[i] = allpar[i].replace('1_T1', 'lambda1')
-        
-        for i, _ in enumerate(eqn):
-            while '1_T1' in eqn[i]:
-                eqn[i] = eqn[i].replace('1_T1', 'lambda1')
-                
-        # make constrained functions
-        cgen= CstrFnGenerator(alldef, eqn, allpar, self.parnames)
-        
-        # get the functions and initial parameters
-        fit_files = self.bfit.fit_files
-        fetch_files = self.bfit.fetch_files
-        fitfns = []
-        par = []
-        rebin = []
-        omit = []
-        fnptrs = []
-        constr_fns = []
-        
-        keylist = sorted(fit_files.fit_lines.keys())
-        for k in keylist:
-            line = fetch_files.data_lines[k]
-            data = line.bdfit
-            
-            # get pulse length
-            pulse_len = -1
-            try:
-                pulse_len = data.bd.pulse_s
-            except (KeyError, AttributeError):
-                pass
-            
-            # get function
-            fn = fit_files.fitter.get_fn(fn_name=fit_files.fit_function_title.get(), 
-                                         ncomp=fit_files.n_component.get(), 
-                                         pulse_len=pulse_len, 
-                                         lifetime=bd.life[fit_files.probe_label['text']])
-            
-            genf, genc = cgen(data=data, fn=fn)
-            fitfns.append(genf)
-            fnptrs.append(fn)
-            constr_fns.append(genc)
-            
-            # get initial parameters
-            par.append(data.fitpar)
-            
-            # get rebin
-            rebin.append(data.rebin.get())
-            
-            # get bin omission
-            omit.append(data.omit.get())
-        
-        # clean up omit strings
-        for i, om in enumerate(omit):
-            if om == fetch_files.bin_remove_starter_line:
-                omit[i] = ''
-        
-        # set up p0, bounds
-        p0 = self.new_par['p0'].values
-        blo = self.new_par['blo'].values
-        bhi = self.new_par['bhi'].values
-        
-        p0 = [[p]*len(keylist) for p in p0]
-        blo = [[p]*len(keylist) for p in blo]
-        bhi = [[p]*len(keylist) for p in bhi]
-                
-        for n in sorted(self.parnames):
-            if n not in defined:
-                p0.append( [p['p0' ][n] for p in par])
[…]
