### 2021-09-28 845078b
Fixed more issues with fit constr
