### 2021-09-28 3ef1152
Added constraint equations to export header
