+++ bfit/gui/tab_fit_files.py
@@ -1140,10 +1140,21 @@ class fit_files(object):
                           '# Number of components: %d' % data.ncomp,
                           '# Global Chi-Squared: %s' % self.gchi_label['text']
                           ]
+                          
+                # add constrained equations to header
+                if self.pop_fitconstr.defined:
+                    head2 = ['# ',
+                             '# Constrained parameters',
+                             ]
+                    header.extend(head2)
+                    
+                    for d, e in zip(self.pop_fitconstr.defined, self.pop_fitconstr.eqn):
+                        header.append('# {defi} = {eqn}'.format(defi=d, eqn=e))
+                    
             else:
                 header = []
 
-            header.extend(['# Generated by bfit v%s on %s' % (__version__, datetime.datetime.now()),
+            header.extend(['#\n# Generated by bfit v%s on %s' % (__version__, datetime.datetime.now()),
                           '#\n#\n'])
 
             with open(filename, 'w') as fid:
