### 2021-09-28 2bf1f41
Fixed syncing issue with changing ncomp
