### 2021-09-28 570cc81
Fixed constrained show components
