+++ bfit/gui/tab_fit_files.py
@@ -1473,6 +1473,7 @@ class fitline(object):
             data            fitdata object in bfit.data dictionary
             disable_entry_callback  disables copy of entry strings to
                                     dataline.bdfit parameter values
+            gui_param_button ttk.Button, set initial parameters 
             lines           list of InputLine objects
             parent          pointer to parent object (frame)
             run_label       label for showing which run is selected
@@ -1558,6 +1559,7 @@ class fitline(object):
 
         # save frame
         self.fitframe = fitframe
+        self.gui_param_button = gui_param_button
 
         # resizing
         for i in range(c):
@@ -1675,25 +1677,21 @@ class fitline(object):
 
         # get number of components and parameter names
         ncomp = fit_files.n_component.get()
-        pnames_single = fit_files.fitter.gen_param_names(fn_name, 1, bdfit.constrained)
-        pnames_combined = fit_files.fitter.gen_param_names(fn_name, ncomp, bdfit.constrained)
+        pnames_single = fit_files.fitter.gen_param_names(fn_name, 1)
+        pnames_combined = fit_files.fitter.gen_param_names(fn_name, ncomp)
 
         if '2' in bdfit.mode:
             fn_single = fit_files.fitter.get_fn(fn_name=fn_name, 
                             ncomp=1,
                             pulse_len=bdfit.pulse_s,
-                            lifetime=bd.life[bfit.probe_species.get()],
-                            constr=bdfit.constrained)
+                            lifetime=bd.life[bfit.probe_species.get()])
             fn_combined = fit_files.fitter.get_fn(fn_name=fn_name, 
                             ncomp=ncomp,
                             pulse_len=bdfit.pulse_s,
-                            lifetime=bd.life[bfit.probe_species.get()],
-                            constr=bdfit.constrained)
+                            lifetime=bd.life[bfit.probe_species.get()])
         else:
-            fn_single = fit_files.fitter.get_fn(fn_name=fn_name, ncomp=1, 
-                                                constr=bdfit.constrained)
-            fn_combined = fit_files.fitter.get_fn(fn_name=fn_name, ncomp=ncomp, 
-                                                constr=bdfit.constrained)
+            fn_single = fit_files.fitter.get_fn(fn_name=fn_name, ncomp=1)
+            fn_combined = fit_files.fitter.get_fn(fn_name=fn_name, ncomp=ncomp)
 
         # draw in redraw mode
         draw_mode = bfit.draw_style.get()
@@ -1871,8 +1869,11 @@ class fitline(object):
             self.bfit.fit_files.pop_fitconstr.disable_constrained_par()
         
             # set parameters again - ensure constr param synced
-            for i, k in enumerate(param_values.index):
-                self.lines[i].set(k, **param_values.loc[k].to_dict())
+            try:
+                for i, k in enumerate(param_values.index):
+                    self.lines[i].set(k, **param_values.loc[k].to_dict())
+            except IndexError:
+                pass
         
     # ======================================================================= #
     def set(self, pname, **kwargs):
