### 2021-09-28 dd25538
Disables initial value button if constrained
