### 2021-09-28 7d5641a
Fixed overly persistent constraints
